refactor(backend): modularize archive maintenance
