+++ docs/specs/README.md
@@ -165,3 +165,4 @@
 | 7gb5w | 号池分组设置弹窗“绑定代理节点”目录加载与同步热修 | 已完成 | `7gb5w-account-pool-bound-proxy-dialog-freshness/SPEC.md` | 2026-04-12 | fast-track / account-pool group settings dialog bound-proxy catalog tri-state + loading placeholder + settings invalidation sync / review-loop clear / local vitest+build+storybook green / screenshot approved / PR #335 merge-ready |
 | ts6qp | account-pool 前端类型债清偿 | 已完成 | `ts6qp-account-pool-typing-debt/SPEC.md` | 2026-04-12 | normal-flow / production `@ts-nocheck` removed from account-pool pages / `UpstreamAccounts` helper extraction + Create strict typing landed / local lint+test+build+storybook green / visual evidence archived |
 | 9vau7 | 后端结构债双 PR 快车道 | 已完成 | `9vau7-backend-structure-dual-pr-followup/SPEC.md` | 2026-04-12 | fast-track / PR1=`proxy+upstream_accounts/routing` 已合并 / PR2=`api+router builders+api-read smoke` 本地 cargo fmt+check+test 与 shared-testbox api-read+proxy-parallel+raw smoke 通过 / merge+cleanup |
+| g7n33 | 后端 archive 结构收敛 | 已实现，待 PR / CI / review-proof 收敛 | `g7n33-backend-archive-structure-convergence/SPEC.md` | 2026-04-12 | fast-track / archive-only follow-up / split `src/maintenance/archive.rs` into real submodules / local cargo fmt+check+test + shared-testbox raw/api-read smoke green / merge+cleanup |
