+++ src/maintenance/archive.rs
@@ -1,3103 +1,13 @@
-async fn backfill_invocation_archive_expiries(
-    pool: &Pool<Sqlite>,
-    config: &AppConfig,
-) -> Result<usize> {
-    let candidates = sqlx::query_as::<_, ArchiveExpiryBackfillCandidate>(
-        r#"
-        SELECT id, coverage_end_at
-        FROM archive_batches
-        WHERE dataset = 'codex_invocations'
-          AND status = ?1
-          AND coverage_end_at IS NOT NULL
-          AND archive_expires_at IS NULL
-          AND historical_rollups_materialized_at IS NOT NULL
-        "#,
-    )
-    .bind(ARCHIVE_STATUS_COMPLETED)
-    .fetch_all(pool)
-    .await?;
-
-    let mut updated = 0usize;
-    for candidate in candidates {
-        let archive_expires_at = shanghai_archive_expiry_from_reference_timestamp(
-            &candidate.coverage_end_at,
-            config.invocation_archive_ttl_days,
-        )?;
-        sqlx::query("UPDATE archive_batches SET archive_expires_at = ?1 WHERE id = ?2")
-            .bind(archive_expires_at)
-            .bind(candidate.id)
-            .execute(pool)
-            .await?;
-        updated += 1;
-    }
-    Ok(updated)
-}
-
-fn classify_raw_compression_alert(
-    oldest_uncompressed_age_secs: u64,
-    uncompressed_bytes: u64,
-) -> RawCompressionAlertLevel {
-    const GIB: u64 = 1024 * 1024 * 1024;
-    if oldest_uncompressed_age_secs >= 48 * 3600 || uncompressed_bytes >= 20 * GIB {
-        RawCompressionAlertLevel::Critical
-    } else if oldest_uncompressed_age_secs >= 24 * 3600 || uncompressed_bytes >= 10 * GIB {
-        RawCompressionAlertLevel::Warn
-    } else {
-        RawCompressionAlertLevel::Ok
-    }
-}
-
-fn invocation_status_counts_toward_terminal_totals(status: Option<&str>) -> bool {
-    let normalized_status = status.map(str::trim).unwrap_or_default();
-    !normalized_status.eq_ignore_ascii_case("running")
-        && !normalized_status.eq_ignore_ascii_case("pending")
-}
-
-async fn load_raw_compression_backlog_snapshot(
-    pool: &Pool<Sqlite>,
-    config: &AppConfig,
-) -> Result<RawCompressionBacklogSnapshot> {
-    let cutoff = shanghai_local_cutoff_for_age_secs_string(config.proxy_raw_hot_secs);
-    let prune_cutoff = shanghai_local_cutoff_string(config.invocation_success_full_days);
-    let archive_cutoff = shanghai_local_cutoff_string(config.invocation_max_days);
-    let row = sqlx::query_as::<_, RawCompressionBacklogAggRow>(
-        r#"
-        SELECT
-            COUNT(*) AS uncompressed_count,
-            COALESCE(SUM(raw_size), 0) AS uncompressed_bytes,
-            MIN(occurred_at) AS oldest_occurred_at
-        FROM (
-            SELECT occurred_at, COALESCE(request_raw_size, 0) AS raw_size
-            FROM codex_invocations
-            WHERE occurred_at < ?1
-              AND occurred_at >= ?2
-              AND (
-                status != 'success'
-                OR detail_level IS NULL
-                OR detail_level != ?3
-                OR occurred_at >= ?4
-              )
-              AND request_raw_path IS NOT NULL
-              AND request_raw_codec = 'identity'
-            UNION ALL
-            SELECT occurred_at, COALESCE(response_raw_size, 0) AS raw_size
-            FROM codex_invocations
-            WHERE occurred_at < ?1
-              AND occurred_at >= ?2
-              AND (
-                status != 'success'
-                OR detail_level IS NULL
-                OR detail_level != ?3
-                OR occurred_at >= ?4
-              )
-              AND response_raw_path IS NOT NULL
-              AND response_raw_codec = 'identity'
-        )
-        "#,
-    )
-    .bind(&cutoff)
-    .bind(&archive_cutoff)
-    .bind(DETAIL_LEVEL_FULL)
-    .bind(&prune_cutoff)
-    .fetch_one(pool)
-    .await?;
-
-    let oldest_uncompressed_age_secs = row
-        .oldest_occurred_at
-        .as_deref()
-        .map(parse_shanghai_local_naive)
-        .transpose()?
-        .map(|oldest| {
-            let now = Utc::now().with_timezone(&Shanghai).naive_local();
-            now.signed_duration_since(oldest).num_seconds().max(0) as u64
-        })
-        .unwrap_or_default();
-    let uncompressed_count = row.uncompressed_count.max(0) as u64;
-    let uncompressed_bytes = row.uncompressed_bytes.unwrap_or_default().max(0) as u64;
-    let alert_level =
-        classify_raw_compression_alert(oldest_uncompressed_age_secs, uncompressed_bytes);
-    Ok(RawCompressionBacklogSnapshot {
-        oldest_uncompressed_age_secs,
-        uncompressed_count,
-        uncompressed_bytes,
-        alert_level,
-    })
-}
-
-async fn log_raw_compression_backlog_if_needed(
-    pool: &Pool<Sqlite>,
-    config: &AppConfig,
-) -> Result<()> {
-    let snapshot = load_raw_compression_backlog_snapshot(pool, config).await?;
-    match snapshot.alert_level {
-        RawCompressionAlertLevel::Ok => {}
-        RawCompressionAlertLevel::Warn => {
-            warn!(
-                oldest_uncompressed_age_secs = snapshot.oldest_uncompressed_age_secs,
-                uncompressed_count = snapshot.uncompressed_count,
-                uncompressed_bytes = snapshot.uncompressed_bytes,
-                alert_level = "warn",
-                "raw compression backlog is above warning threshold"
-            );
-        }
-        RawCompressionAlertLevel::Critical => {
-            error!(
-                oldest_uncompressed_age_secs = snapshot.oldest_uncompressed_age_secs,
-                uncompressed_count = snapshot.uncompressed_count,
-                uncompressed_bytes = snapshot.uncompressed_bytes,
-                alert_level = "critical",
-                "raw compression backlog is above critical threshold"
-            );
-        }
-    }
-    Ok(())
-}
-
-fn archive_file_is_stale_temp(path: &Path) -> bool {
-    path.file_name()
-        .and_then(|value| value.to_str())
-        .map(is_archive_temp_file_name)
-        .unwrap_or(false)
-}
-
-fn archive_temp_file_is_old_enough(path: &Path) -> bool {
-    fs::metadata(path)
-        .ok()
-        .and_then(|metadata| metadata.modified().ok())
-        .and_then(|modified| modified.elapsed().ok())
-        .map(|age| age.as_secs() >= DEFAULT_ARCHIVE_TEMP_MIN_AGE_SECS)
-        .unwrap_or(false)
-}
-
-fn archive_file_size(path: &Path) -> u64 {
-    fs::metadata(path)
-        .map(|metadata| metadata.len())
-        .unwrap_or_default()
-}
-
-fn cleanup_stale_archive_temp_files(
-    config: &AppConfig,
-    dry_run: bool,
-) -> Result<ArchiveTempCleanupSummary> {
-    let archive_root = resolved_archive_dir(config);
-    let mut files = Vec::new();
-    collect_archive_file_paths(&archive_root, &mut files)?;
-    let mut summary = ArchiveTempCleanupSummary::default();
-    for file_path in files {
-        if !archive_file_is_stale_temp(&file_path) || !archive_temp_file_is_old_enough(&file_path) {
-            continue;
-        }
-        let file_size = archive_file_size(&file_path);
-        if dry_run {
-            summary.stale_temp_files_removed += 1;
-            summary.stale_temp_bytes_removed += file_size;
-            continue;
-        }
-        match fs::remove_file(&file_path) {
-            Ok(_) => {
-                summary.stale_temp_files_removed += 1;
-                summary.stale_temp_bytes_removed += file_size;
-            }
-            Err(err) if err.kind() == io::ErrorKind::NotFound => {}
-            Err(err) => {
-                warn!(
-                    file_path = %file_path.display(),
-                    error = %err,
-                    "failed to remove stale archive temp file"
-                );
-            }
-        }
-    }
-    Ok(summary)
-}
-
-async fn verify_archive_storage(
-    pool: &Pool<Sqlite>,
-    config: &AppConfig,
-) -> Result<ArchiveStorageVerificationSummary> {
-    let manifest_rows = sqlx::query_as::<_, ArchiveStorageManifestRow>(
-        r#"
-        SELECT id, dataset, layout, file_path
-        FROM archive_batches
-        WHERE status = ?1
-        "#,
-    )
-    .bind(ARCHIVE_STATUS_COMPLETED)
-    .fetch_all(pool)
-    .await?;
-    let mut summary = ArchiveStorageVerificationSummary {
-        manifest_rows: manifest_rows.len(),
-        ..ArchiveStorageVerificationSummary::default()
-    };
-    let referenced_paths = manifest_rows
-        .iter()
-        .map(|row| normalize_path_for_compare(Path::new(&row.file_path)))
-        .collect::<HashSet<_>>();
-    for row in &manifest_rows {
-        if !Path::new(&row.file_path).exists() {
-            summary.missing_files += 1;
-            warn!(
-                archive_batch_id = row.id,
-                dataset = row.dataset,
-                layout = row.layout,
-                file_path = row.file_path,
-                "archive manifest points to a missing file"
-            );
-        }
-    }
-
-    let archive_root = resolved_archive_dir(config);
-    let mut files = Vec::new();
-    collect_archive_file_paths(&archive_root, &mut files)?;
-    for file_path in files {
-        let normalized = normalize_path_for_compare(&file_path);
-        if archive_file_is_stale_temp(&file_path) {
-            summary.stale_temp_files += 1;
-            summary.stale_temp_bytes += archive_file_size(&file_path);
-            continue;
-        }
-        if !referenced_paths.contains(&normalized) {
-            summary.orphan_files += 1;
-        }
-    }
-    Ok(summary)
-}
-
-#[derive(Debug, FromRow)]
-struct ArchiveBatchCleanupCandidate {
-    id: i64,
-    dataset: String,
-    file_path: String,
-    historical_rollups_materialized_at: Option<String>,
-    coverage_end_at: Option<String>,
-}
-
-async fn cleanup_expired_archive_batches(
-    pool: &Pool<Sqlite>,
-    config: &AppConfig,
-    dry_run: bool,
-) -> Result<usize> {
-    if !dry_run {
-        backfill_invocation_archive_expiries(pool, config).await?;
-    }
-    let cutoff = format_naive(Utc::now().with_timezone(&Shanghai).naive_local());
-    let invocation_archive_cutoff = shanghai_local_cutoff_string(config.invocation_max_days);
-    let candidates = sqlx::query_as::<_, ArchiveBatchCleanupCandidate>(
-        r#"
-        SELECT id, dataset, file_path, historical_rollups_materialized_at, coverage_end_at
-        FROM archive_batches
-        WHERE status = ?1
-          AND archive_expires_at IS NOT NULL
-          AND archive_expires_at < ?2
-        ORDER BY archive_expires_at ASC, id ASC
-        "#,
-    )
-    .bind(ARCHIVE_STATUS_COMPLETED)
-    .bind(&cutoff)
-    .fetch_all(pool)
-    .await?;
-
-    let mut eligible_candidates = Vec::new();
-    for candidate in candidates {
-        if HISTORICAL_ROLLUP_ARCHIVE_DATASETS.contains(&candidate.dataset.as_str())
-            && candidate.historical_rollups_materialized_at.is_none()
-        {
-            continue;
-        }
-        if candidate.dataset == HOURLY_ROLLUP_DATASET_INVOCATIONS {
-            if candidate
-                .coverage_end_at
-                .as_deref()
-                .map(|coverage_end_at| coverage_end_at >= invocation_archive_cutoff.as_str())
-                .unwrap_or(true)
-            {
-                continue;
-            }
-        }
-        eligible_candidates.push(candidate);
-    }
-
-    if dry_run {
-        for candidate in &eligible_candidates {
-            info!(
-                dataset = candidate.dataset,
-                file_path = candidate.file_path,
-                "retention dry-run planned archive batch cleanup"
-            );
-        }
-        return Ok(eligible_candidates.len());
-    }
-
-    let mut deleted = 0usize;
-    for candidate in eligible_candidates {
-        match fs::remove_file(&candidate.file_path) {
-            Ok(_) => {}
-            Err(err) if err.kind() == io::ErrorKind::NotFound => {}
-            Err(err) => {
-                warn!(
-                    dataset = candidate.dataset,
-                    file_path = candidate.file_path,
-                    error = %err,
-                    "failed to remove expired archive batch file; keeping manifest"
-                );
-                continue;
-            }
-        }
-
-        let mut tx = pool.begin().await?;
-        sqlx::query("DELETE FROM archive_batch_upstream_activity WHERE archive_batch_id = ?1")
-            .bind(candidate.id)
-            .execute(tx.as_mut())
-            .await?;
-        sqlx::query("DELETE FROM archive_batches WHERE id = ?1")
-            .bind(candidate.id)
-            .execute(tx.as_mut())
-            .await?;
-        sqlx::query(
-            "DELETE FROM hourly_rollup_archive_replay WHERE dataset = ?1 AND file_path = ?2",
-        )
-        .bind(&candidate.dataset)
-        .bind(&candidate.file_path)
-        .execute(tx.as_mut())
-        .await?;
-        tx.commit().await?;
-        deleted += 1;
-    }
-
-    Ok(deleted)
-}
-
-#[derive(Debug, FromRow)]
-struct HistoricalRollupPendingArchiveBatchRow {
-    dataset: String,
-    month_key: String,
-    file_path: String,
-    coverage_start_at: Option<String>,
-    coverage_end_at: Option<String>,
-}
-
-#[derive(Debug, FromRow)]
-struct LegacyArchivePruneCandidateRow {
-    id: i64,
-    dataset: String,
-    file_path: String,
-    historical_rollups_materialized_at: Option<String>,
-    coverage_end_at: Option<String>,
-}
-
-fn estimate_historical_rollup_pending_bucket_count(
-    row: &HistoricalRollupPendingArchiveBatchRow,
-) -> u64 {
-    if let (Some(start), Some(end)) = (&row.coverage_start_at, &row.coverage_end_at)
-        && let (Ok(start_local), Ok(end_local)) = (
-            parse_shanghai_local_naive(start),
-            parse_shanghai_local_naive(end),
-        )
-    {
-        let start_utc = local_naive_to_utc(start_local, Shanghai);
-        let end_utc = local_naive_to_utc(end_local, Shanghai);
-        let secs = (end_utc.timestamp() - start_utc.timestamp()).max(0);
-        return ((secs + 3_599) / 3_600).max(1) as u64;
-    }
-
-    let Ok(start_date) = NaiveDate::parse_from_str(&format!("{}-01", row.month_key), "%Y-%m-%d")
-    else {
-        return 0;
-    };
-    let (next_year, next_month) = if start_date.month() == 12 {
-        (start_date.year() + 1, 1)
-    } else {
-        (start_date.year(), start_date.month() + 1)
-    };
-    let Some(next_month_date) = NaiveDate::from_ymd_opt(next_year, next_month, 1) else {
-        return 0;
-    };
-    let Some(start_naive) = start_date.and_hms_opt(0, 0, 0) else {
-        return 0;
-    };
-    let Some(end_naive) = next_month_date.and_hms_opt(0, 0, 0) else {
-        return 0;
-    };
-    let start_utc = local_naive_to_utc(start_naive, Shanghai);
-    let end_utc = local_naive_to_utc(end_naive, Shanghai);
-    ((end_utc.timestamp() - start_utc.timestamp()).max(0) / 3_600) as u64
-}
-
-async fn count_historical_rollup_archive_batches(
-    pool: &Pool<Sqlite>,
-    pending_only: bool,
-) -> Result<i64> {
-    let mut query =
-        QueryBuilder::<Sqlite>::new("SELECT COUNT(*) FROM archive_batches WHERE status = ");
-    query.push_bind(ARCHIVE_STATUS_COMPLETED);
-    query.push(" AND dataset IN (");
-    {
-        let mut separated = query.separated(", ");
-        for dataset in HISTORICAL_ROLLUP_ARCHIVE_DATASETS {
-            separated.push_bind(dataset);
-        }
-    }
-    query.push(")");
-    if pending_only {
-        query.push(" AND historical_rollups_materialized_at IS NULL");
-    }
-    Ok(query.build_query_scalar::<i64>().fetch_one(pool).await?)
-}
-
-fn historical_rollup_materialized_bucket_targets() -> [&'static str; 7] {
-    [
-        HOURLY_ROLLUP_TARGET_INVOCATIONS,
-        HOURLY_ROLLUP_TARGET_INVOCATION_FAILURES,
-        HOURLY_ROLLUP_TARGET_PROXY_PERF,
-        HOURLY_ROLLUP_TARGET_PROMPT_CACHE,
-        HOURLY_ROLLUP_TARGET_PROMPT_CACHE_UPSTREAM_ACCOUNTS,
-        HOURLY_ROLLUP_TARGET_STICKY_KEYS,
-        HOURLY_ROLLUP_TARGET_FORWARD_PROXY_ATTEMPTS,
-    ]
-}
-
-async fn load_latest_materialized_legacy_invocation_rollup_bucket_epoch(
-    pool: &Pool<Sqlite>,
-    config: &AppConfig,
-) -> Result<Option<i64>> {
-    let invocation_archive_cutoff = shanghai_local_cutoff_string(config.invocation_max_days);
-    let latest_coverage_end_at: Option<String> = sqlx::query_scalar(
-        r#"
-        SELECT MAX(coverage_end_at)
-        FROM archive_batches
-        WHERE dataset = 'codex_invocations'
-          AND status = ?1
-          AND historical_rollups_materialized_at IS NOT NULL
-          AND coverage_end_at IS NOT NULL
-          AND coverage_end_at < ?2
-        "#,
-    )
-    .bind(ARCHIVE_STATUS_COMPLETED)
-    .bind(invocation_archive_cutoff)
-    .fetch_one(pool)
-    .await?;
-
-    Ok(latest_coverage_end_at.and_then(|coverage_end_at| {
-        parse_shanghai_local_naive(&coverage_end_at)
-            .ok()
-            .and_then(|naive| {
-                let bucket_start_epoch =
-                    align_bucket_epoch(local_naive_to_utc(naive, Shanghai).timestamp(), 3_600, 0);
-                Utc.timestamp_opt(bucket_start_epoch, 0)
-                    .single()
-                    .map(|_| bucket_start_epoch)
-            })
-    }))
-}
-
-async fn count_materialized_historical_rollup_buckets(pool: &Pool<Sqlite>) -> Result<i64> {
-    let mut query = QueryBuilder::<Sqlite>::new(
-        "SELECT COUNT(*) FROM hourly_rollup_materialized_buckets WHERE target IN (",
-    );
-    {
-        let mut separated = query.separated(", ");
-        for target in historical_rollup_materialized_bucket_targets() {
-            separated.push_bind(target);
-        }
-    }
-    query.push(")");
-    Ok(query.build_query_scalar::<i64>().fetch_one(pool).await?)
-}
-
-pub(crate) async fn load_historical_rollup_backfill_snapshot(
-    pool: &Pool<Sqlite>,
-    config: &AppConfig,
-) -> Result<HistoricalRollupBackfillSnapshot> {
-    let mut query = QueryBuilder::<Sqlite>::new(
-        "SELECT dataset, month_key, file_path, coverage_start_at, coverage_end_at \
-         FROM archive_batches WHERE status = ",
-    );
-    query.push_bind(ARCHIVE_STATUS_COMPLETED);
-    query.push(" AND historical_rollups_materialized_at IS NULL AND dataset IN (");
-    {
-        let mut separated = query.separated(", ");
-        for dataset in HISTORICAL_ROLLUP_ARCHIVE_DATASETS {
-            separated.push_bind(dataset);
-        }
-    }
-    query.push(") ORDER BY month_key ASC, id ASC");
-    let pending_rows = query
-        .build_query_as::<HistoricalRollupPendingArchiveBatchRow>()
-        .fetch_all(pool)
-        .await?;
-    let pending_buckets = pending_rows
-        .iter()
-        .filter(|row| Path::new(&row.file_path).exists())
-        .map(estimate_historical_rollup_pending_bucket_count)
-        .sum::<u64>();
-    let legacy_archive_pending = pending_rows
-        .iter()
-        .filter(|row| Path::new(&row.file_path).exists())
-        .count() as u64;
-    let legacy_invocation_pending = pending_rows
-        .iter()
-        .filter(|row| Path::new(&row.file_path).exists())
-        .any(|row| row.dataset == HOURLY_ROLLUP_DATASET_INVOCATIONS);
-    let last_materialized_hour =
-        load_latest_materialized_legacy_invocation_rollup_bucket_epoch(pool, config)
-            .await?
-            .and_then(|epoch| Utc.timestamp_opt(epoch, 0).single())
-            .map(format_utc_iso);
-    let alert_level = if legacy_archive_pending == 0 {
-        HistoricalRollupBackfillAlertLevel::None
-    } else if legacy_invocation_pending {
-        HistoricalRollupBackfillAlertLevel::Critical
-    } else {
-        HistoricalRollupBackfillAlertLevel::Warn
-    };
-
-    Ok(HistoricalRollupBackfillSnapshot {
-        pending_buckets,
-        legacy_archive_pending,
-        last_materialized_hour,
-        alert_level,
-    })
-}
-
-async fn materialize_historical_rollups(
-    pool: &Pool<Sqlite>,
-    config: &AppConfig,
-    dry_run: bool,
-) -> Result<HistoricalRollupMaterializationSummary> {
-    let scanned_archive_batches = count_historical_rollup_archive_batches(pool, false).await?;
-    let pending_snapshot = load_historical_rollup_backfill_snapshot(pool, config).await?;
-    if dry_run {
-        return Ok(HistoricalRollupMaterializationSummary {
-            scanned_archive_batches: scanned_archive_batches as usize,
-            materialized_archive_batches: pending_snapshot.legacy_archive_pending as usize,
-            materialized_bucket_count: pending_snapshot.pending_buckets as usize,
-            materialized_invocation_batches: 0,
-            materialized_forward_proxy_batches: 0,
-            last_materialized_bucket_start_epoch:
-                load_latest_materialized_legacy_invocation_rollup_bucket_epoch(pool, config).await?,
-        });
-    }
-
-    let mut tx = pool.begin().await?;
-    let materialized_invocation_batches =
-        replay_invocation_archives_into_hourly_rollups_tx(tx.as_mut()).await?;
-    let materialized_forward_proxy_batches =
-        replay_forward_proxy_archives_into_hourly_rollups_tx(tx.as_mut()).await?;
-    loop {
-        let updated = replay_live_invocation_hourly_rollups_tx(tx.as_mut()).await?;
-        if updated == 0 {
-            break;
-        }
-    }
-    loop {
-        let updated = replay_live_forward_proxy_attempt_hourly_rollups_tx(tx.as_mut()).await?;
-        if updated == 0 {
-            break;
-        }
-    }
-    tx.commit().await?;
-
-    Ok(HistoricalRollupMaterializationSummary {
-        scanned_archive_batches: scanned_archive_batches as usize,
-        materialized_archive_batches: (materialized_invocation_batches
-            + materialized_forward_proxy_batches) as usize,
-        materialized_bucket_count: count_materialized_historical_rollup_buckets(pool).await?
-            as usize,
-        materialized_invocation_batches: materialized_invocation_batches as usize,
-        materialized_forward_proxy_batches: materialized_forward_proxy_batches as usize,
-        last_materialized_bucket_start_epoch:
-            load_latest_materialized_legacy_invocation_rollup_bucket_epoch(pool, config).await?,
-    })
-}
-
-async fn prune_legacy_archive_batches(
-    pool: &Pool<Sqlite>,
-    config: &AppConfig,
-    dry_run: bool,
-) -> Result<LegacyArchivePruneSummary> {
-    let mut query = QueryBuilder::<Sqlite>::new(
-        "SELECT id, dataset, file_path, historical_rollups_materialized_at, coverage_end_at \
-         FROM archive_batches WHERE status = ",
-    );
-    query.push_bind(ARCHIVE_STATUS_COMPLETED);
-    query.push(" AND dataset IN (");
-    {
-        let mut separated = query.separated(", ");
-        for dataset in HISTORICAL_ROLLUP_ARCHIVE_DATASETS {
-            separated.push_bind(dataset);
-        }
-    }
-    query.push(") AND COALESCE(layout, ");
-    query.push_bind(ARCHIVE_LAYOUT_LEGACY_MONTH);
-    query.push(") = ");
-    query.push_bind(ARCHIVE_LAYOUT_LEGACY_MONTH);
-    query.push(" ORDER BY month_key ASC, id ASC");
-    let candidates = query
-        .build_query_as::<LegacyArchivePruneCandidateRow>()
-        .fetch_all(pool)
-        .await?;
-
-    let pending_account_count = count_upstream_accounts_missing_last_activity(pool).await?;
-    let invocation_archive_cutoff = shanghai_local_cutoff_string(config.invocation_max_days);
-    let mut summary = LegacyArchivePruneSummary {
-        scanned_archive_batches: candidates.len(),
-        ..LegacyArchivePruneSummary::default()
-    };
-
-    for candidate in candidates {
-        let file_missing = !Path::new(&candidate.file_path).exists();
-
-        if candidate.dataset == HOURLY_ROLLUP_DATASET_INVOCATIONS && pending_account_count > 0 {
-            summary.skipped_unmaterialized_batches += 1;
-            continue;
-        }
-
-        if !file_missing && candidate.historical_rollups_materialized_at.is_none() {
-            summary.skipped_unmaterialized_batches += 1;
-            continue;
-        }
-
-        if candidate.dataset == HOURLY_ROLLUP_DATASET_INVOCATIONS
-            && candidate
-                .coverage_end_at
-                .as_deref()
-                .map(|coverage_end_at| coverage_end_at >= invocation_archive_cutoff.as_str())
-                .unwrap_or(true)
-        {
-            summary.skipped_retained_batches += 1;
-            continue;
-        }
-
-        if dry_run {
-            info!(
-                dataset = candidate.dataset,
-                file_path = candidate.file_path,
-                "maintenance dry-run planned legacy archive prune"
-            );
-            summary.deleted_archive_batches += 1;
-            continue;
-        }
-
-        match fs::remove_file(&candidate.file_path) {
-            Ok(_) => {}
-            Err(err) if err.kind() == io::ErrorKind::NotFound => {}
-            Err(err) => {
-                warn!(
-                    dataset = candidate.dataset,
-                    file_path = candidate.file_path,
-                    error = %err,
-                    "failed to remove legacy archive batch file; keeping metadata"
-                );
-                summary.skipped_unmaterialized_batches += 1;
-                continue;
-            }
-        }
-
-        let mut tx = pool.begin().await?;
-        sqlx::query("DELETE FROM archive_batch_upstream_activity WHERE archive_batch_id = ?1")
-            .bind(candidate.id)
-            .execute(tx.as_mut())
-            .await?;
-        sqlx::query(
-            "DELETE FROM hourly_rollup_archive_replay WHERE dataset = ?1 AND file_path = ?2",
-        )
-        .bind(&candidate.dataset)
-        .bind(&candidate.file_path)
-        .execute(tx.as_mut())
-        .await?;
-        sqlx::query("DELETE FROM archive_batches WHERE id = ?1")
-            .bind(candidate.id)
-            .execute(tx.as_mut())
-            .await?;
-        tx.commit().await?;
-        summary.deleted_archive_batches += 1;
-    }
-
-    Ok(summary)
-}
-
-async fn prune_archive_batches(
-    pool: &Pool<Sqlite>,
-    config: &AppConfig,
-    dry_run: bool,
-) -> Result<ArchiveBatchPruneSummary> {
-    let expired_archive_batches_deleted =
-        cleanup_expired_archive_batches(pool, config, dry_run).await?;
-    let legacy_summary = prune_legacy_archive_batches(pool, config, dry_run).await?;
-    Ok(ArchiveBatchPruneSummary {
-        expired_archive_batches_deleted,
-        legacy_archive_batches_deleted: legacy_summary.deleted_archive_batches,
-    })
-}
-
-async fn compact_old_quota_snapshots(
-    pool: &Pool<Sqlite>,
-    config: &AppConfig,
-    dry_run: bool,
-) -> Result<(usize, usize)> {
-    let cutoff = shanghai_utc_cutoff_string(config.quota_snapshot_full_days);
-    let spec = archive_table_spec("codex_quota_snapshots");
-
-    if dry_run {
-        let batch_counts = sqlx::query_as::<_, DryRunBatchCount>(
-            r#"
-            WITH ranked AS (
-                SELECT
-                    captured_at,
-                    ROW_NUMBER() OVER (
-                        PARTITION BY strftime('%Y-%m-%d', datetime(captured_at, '+8 hours'))
-                        ORDER BY captured_at DESC, id DESC
-                    ) AS row_num
-                FROM codex_quota_snapshots
-                WHERE captured_at < ?1
-            )
-            SELECT strftime('%Y-%m', datetime(captured_at, '+8 hours')) AS month_key,
-                   COUNT(*) AS row_count
-            FROM ranked
-            WHERE row_num > 1
-            GROUP BY 1
-            ORDER BY 1
-            "#,
-        )
-        .bind(&cutoff)
-        .fetch_all(pool)
-        .await?;
-        for batch in &batch_counts {
-            info!(
-                dataset = spec.dataset,
-                month_key = %batch.month_key,
-                rows = batch.row_count,
-                "retention dry-run planned quota compaction batch"
-            );
-        }
-        return Ok((
-            batch_counts
-                .iter()
-                .map(|batch| batch.row_count as usize)
-                .sum(),
-            batch_counts.len(),
-        ));
-    }
-
-    let mut rows_archived = 0usize;
-    let mut archive_batches = 0usize;
-
-    loop {
-        let candidates = sqlx::query_as::<_, TimestampedArchiveCandidate>(
-            r#"
-            WITH ranked AS (
-                SELECT
-                    id,
-                    captured_at AS timestamp_value,
-                    ROW_NUMBER() OVER (
-                        PARTITION BY strftime('%Y-%m-%d', datetime(captured_at, '+8 hours'))
-                        ORDER BY captured_at DESC, id DESC
-                    ) AS row_num
-                FROM codex_quota_snapshots
-                WHERE captured_at < ?1
-            )
-            SELECT id, timestamp_value
-            FROM ranked
-            WHERE row_num > 1
-            ORDER BY timestamp_value ASC, id ASC
-            LIMIT ?2
-            "#,
-        )
-        .bind(&cutoff)
-        .bind(config.retention_batch_rows as i64)
-        .fetch_all(pool)
-        .await?;
-
-        if candidates.is_empty() {
-            break;
-        }
-
-        let mut by_month: BTreeMap<String, Vec<TimestampedArchiveCandidate>> = BTreeMap::new();
-        for candidate in candidates {
-            let month_key = shanghai_month_key_from_utc_naive(&candidate.timestamp_value)?;
-            by_month.entry(month_key).or_default().push(candidate);
-        }
-
-        for (month_key, group) in by_month {
-            rows_archived += group.len();
-            archive_batches += 1;
-            let ids = group
-                .iter()
-                .map(|candidate| candidate.id)
-                .collect::<Vec<_>>();
-            let mut archive_outcome =
-                archive_rows_into_month_batch(pool, config, spec, &month_key, &ids).await?;
-            set_archive_batch_coverage_from_utc_rows(
-                &mut archive_outcome,
-                group
-                    .iter()
-                    .map(|candidate| candidate.timestamp_value.as_str()),
-            )?;
-            let mut tx = pool.begin().await?;
-            upsert_archive_batch_manifest(tx.as_mut(), &archive_outcome).await?;
-            delete_rows_by_ids(tx.as_mut(), spec.dataset, &ids).await?;
-            tx.commit().await?;
-        }
-    }
-
-    Ok((rows_archived, archive_batches))
-}
-
-async fn refresh_archive_upstream_activity_manifest(
-    pool: &Pool<Sqlite>,
-    dry_run: bool,
-) -> Result<ArchiveManifestRefreshSummary> {
-    let batches = sqlx::query_as::<_, ArchiveManifestBatchRow>(
-        r#"
-        SELECT id, file_path
-        FROM archive_batches
-        WHERE dataset = 'codex_invocations'
-          AND status = ?1
-          AND upstream_activity_manifest_refreshed_at IS NULL
-        ORDER BY month_key DESC, day_key DESC, part_key DESC, created_at DESC, id DESC
-        "#,
-    )
-    .bind(ARCHIVE_STATUS_COMPLETED)
-    .fetch_all(pool)
-    .await?;
-
-    let mut summary = ArchiveManifestRefreshSummary {
-        pending_batches: batches.len(),
-        ..ArchiveManifestRefreshSummary::default()
-    };
-
-    for batch in batches {
-        let archive_path = PathBuf::from(&batch.file_path);
-        if !archive_path.exists() {
-            summary.missing_files += 1;
-            warn!(
-                archive_batch_id = batch.id,
-                file_path = %archive_path.display(),
-                "archive upstream activity manifest rebuild skipped missing archive file and will retry later"
-            );
-            continue;
-        }
-
-        let values = match load_archive_upstream_activity_from_file(&archive_path).await {
-            Ok(values) => values,
-            Err(err) => {
-                warn!(
-                    archive_batch_id = batch.id,
-                    file_path = %archive_path.display(),
-                    error = %err,
-                    "archive upstream activity manifest rebuild failed and will retry later"
-                );
-                continue;
-            }
-        };
-        let deduped_values = dedupe_archive_upstream_last_activity(values);
-        summary.refreshed_batches += 1;
-        summary.account_rows_written += deduped_values.len();
-        if dry_run {
-            continue;
-        }
-
-        let mut tx = pool.begin().await?;
-        write_archive_batch_upstream_activity(tx.as_mut(), batch.id, &deduped_values).await?;
-        tx.commit().await?;
-    }
-
-    Ok(summary)
-}
-
-async fn load_archive_upstream_activity_from_file(
-    archive_path: &Path,
-) -> Result<Vec<(i64, String)>> {
-    let temp_path = PathBuf::from(format!(
-        "{}.{}.sqlite",
-        archive_path.display(),
-        retention_temp_suffix()
-    ));
-    if temp_path.exists() {
-        let _ = fs::remove_file(&temp_path);
-    }
-    let temp_cleanup = TempSqliteCleanup(temp_path.clone());
-    inflate_gzip_sqlite_file(archive_path, &temp_path)?;
-
-    let database_url = format!("sqlite://{}", temp_path.to_string_lossy());
-    let connect_opts = build_sqlite_connect_options(
-        &database_url,
-        Duration::from_secs(DEFAULT_SQLITE_BUSY_TIMEOUT_SECS),
-    )?;
-    let archive_pool = SqlitePoolOptions::new()
-        .max_connections(1)
-        .connect_with(connect_opts)
-        .await
-        .with_context(|| format!("failed to open archive batch {}", archive_path.display()))?;
-
-    let rows = sqlx::query_as::<_, ArchivedAccountLastActivityRow>(
-        r#"
-        SELECT account_id, MAX(occurred_at) AS last_activity_at
-        FROM (
-            SELECT
-                CASE
-                    WHEN json_valid(payload) THEN CAST(json_extract(payload, '$.upstreamAccountId') AS INTEGER)
-                END AS account_id,
-                occurred_at
-            FROM codex_invocations
-        )
-        WHERE account_id IS NOT NULL
-        GROUP BY account_id
-        "#,
-    )
-    .fetch_all(&archive_pool)
-    .await?;
-
-    archive_pool.close().await;
-    drop(temp_cleanup);
-
-    Ok(dedupe_archive_upstream_last_activity(
-        rows.into_iter()
-            .map(|row| (row.account_id, row.last_activity_at)),
-    ))
-}
-
-async fn backfill_upstream_account_last_activity_from_archives(
-    pool: &Pool<Sqlite>,
-    scan_limit: Option<u64>,
-    max_elapsed: Option<Duration>,
-) -> Result<ArchiveBackfillSummary> {
-    let total_pending_accounts = sqlx::query_scalar::<_, i64>(
-        r#"
-        SELECT COUNT(*)
-        FROM pool_upstream_accounts
-        WHERE last_activity_at IS NULL
-          AND last_activity_archive_backfill_completed = 0
-        "#,
-    )
-    .fetch_one(pool)
-    .await?
-    .max(0) as u64;
-    if total_pending_accounts == 0 {
-        return Ok(ArchiveBackfillSummary::default());
-    }
-
-    let pending_fetch_limit = scan_limit.unwrap_or(total_pending_accounts).max(1) as i64;
-    let pending_account_ids = sqlx::query_scalar::<_, i64>(
-        r#"
-        SELECT id
-        FROM pool_upstream_accounts
-        WHERE last_activity_at IS NULL
-          AND last_activity_archive_backfill_completed = 0
-        ORDER BY id ASC
-        LIMIT ?1
-        "#,
-    )
-    .bind(pending_fetch_limit)
-    .fetch_all(pool)
-    .await?;
-
-    let pending_manifest_batches = sqlx::query_scalar::<_, i64>(
-        r#"
-        SELECT COUNT(*)
-        FROM archive_batches
-        WHERE dataset = 'codex_invocations'
-          AND status = ?1
-          AND upstream_activity_manifest_refreshed_at IS NULL
-        "#,
-    )
-    .bind(ARCHIVE_STATUS_COMPLETED)
-    .fetch_one(pool)
-    .await?;
-    if pending_manifest_batches > 0 {
-        return Ok(ArchiveBackfillSummary {
-            waiting_for_manifest_backfill: true,
-            ..ArchiveBackfillSummary::default()
-        });
-    }
-
-    let scanned_batches = sqlx::query_scalar::<_, i64>(
-        r#"
-        SELECT COUNT(*)
-        FROM archive_batches
-        WHERE dataset = 'codex_invocations'
-          AND status = ?1
-        "#,
-    )
-    .bind(ARCHIVE_STATUS_COMPLETED)
-    .fetch_one(pool)
-    .await?
-    .max(0) as u64;
-    if scanned_batches == 0 {
-        sqlx::query(
-            r#"
-            UPDATE pool_upstream_accounts
-            SET last_activity_archive_backfill_completed = 1
-            WHERE last_activity_at IS NULL
-              AND last_activity_archive_backfill_completed = 0
-            "#,
-        )
-        .execute(pool)
-        .await?;
-        return Ok(ArchiveBackfillSummary::default());
-    }
-
-    let pending = pending_account_ids.into_iter().collect::<HashSet<_>>();
-    let mut recovered = HashMap::<i64, String>::new();
-    let pending_chunks = pending_account_ids_chunks(&pending);
-    let started_at = Instant::now();
-    let mut processed_account_ids = HashSet::new();
-    let mut hit_budget = total_pending_accounts > pending.len() as u64;
-
-    for (chunk_idx, account_ids) in pending_chunks.iter().enumerate() {
-        if startup_backfill_budget_reached(
-            started_at,
-            processed_account_ids.len() as u64,
-            None,
-            max_elapsed,
-        ) {
-            hit_budget = true;
-            break;
-        }
-        for account_id in account_ids {
-            processed_account_ids.insert(*account_id);
-        }
-        let mut query = QueryBuilder::<Sqlite>::new(
-            "SELECT account_id, MAX(last_activity_at) AS last_activity_at FROM archive_batch_upstream_activity WHERE account_id IN (",
-        );
-        {
-            let mut separated = query.separated(", ");
-            for account_id in account_ids {
-                separated.push_bind(account_id);
-            }
-        }
-        query.push(") GROUP BY account_id");
-        for row in query
-            .build_query_as::<ArchivedAccountLastActivityRow>()
-            .fetch_all(pool)
-            .await?
-        {
-            recovered
-                .entry(row.account_id)
-                .and_modify(|current| {
-                    if *current < row.last_activity_at {
-                        *current = row.last_activity_at.clone();
-                    }
-                })
-                .or_insert(row.last_activity_at);
-        }
-
-        if chunk_idx + 1 < pending_chunks.len()
-            && startup_backfill_budget_reached(
-                started_at,
-                processed_account_ids.len() as u64,
-                None,
-                max_elapsed,
-            )
-        {
-            hit_budget = true;
-            break;
-        }
-    }
-
-    if recovered.is_empty() {
-        let processed = processed_account_ids.iter().copied().collect::<Vec<_>>();
-        if !processed.is_empty() {
-            mark_archive_backfill_completed_for_accounts(pool, &processed).await?;
-        }
-        return Ok(ArchiveBackfillSummary {
-            scanned_batches: processed_account_ids.len() as u64,
-            updated_accounts: 0,
-            hit_budget,
-            waiting_for_manifest_backfill: false,
-        });
-    }
-
-    let unresolved: Vec<i64> = processed_account_ids
-        .iter()
-        .copied()
-        .filter(|account_id| !recovered.contains_key(account_id))
-        .collect();
-    let updated_accounts = recovered.len() as u64;
-    let mut tx = pool.begin().await?;
-    for (account_id, occurred_at) in recovered {
-        sqlx::query(
-            r#"
-            UPDATE pool_upstream_accounts
-            SET last_activity_at = CASE
-                    WHEN last_activity_at IS NULL OR last_activity_at < ?1 THEN ?1
-                    ELSE last_activity_at
-                END,
-                last_activity_archive_backfill_completed = 1
-            WHERE id = ?2
-            "#,
-        )
-        .bind(occurred_at)
-        .bind(account_id)
-        .execute(tx.as_mut())
-        .await?;
-    }
-    if !unresolved.is_empty() {
-        mark_archive_backfill_completed_for_accounts_tx(tx.as_mut(), &unresolved).await?;
-    }
-    tx.commit().await?;
-
-    Ok(ArchiveBackfillSummary {
-        scanned_batches: processed_account_ids.len() as u64,
-        updated_accounts,
-        hit_budget,
-        waiting_for_manifest_backfill: false,
-    })
-}
-
-fn pending_account_ids_chunks(pending: &HashSet<i64>) -> Vec<Vec<i64>> {
-    pending
-        .iter()
-        .copied()
-        .collect::<Vec<_>>()
-        .chunks(BACKFILL_ACCOUNT_BIND_BATCH_SIZE)
-        .map(|chunk| chunk.to_vec())
-        .collect()
-}
-
-async fn mark_archive_backfill_completed_for_accounts(
-    pool: &Pool<Sqlite>,
-    account_ids: &[i64],
-) -> Result<()> {
-    if account_ids.is_empty() {
-        return Ok(());
-    }
-    for account_chunk in account_ids.chunks(BACKFILL_ACCOUNT_BIND_BATCH_SIZE) {
-        let mut update = QueryBuilder::<Sqlite>::new(
-            "UPDATE pool_upstream_accounts SET last_activity_archive_backfill_completed = 1 WHERE id IN (",
-        );
-        {
-            let mut separated = update.separated(", ");
-            for account_id in account_chunk {
-                separated.push_bind(account_id);
-            }
-        }
-        update.push(")");
-        update.build().execute(pool).await?;
-    }
-    Ok(())
-}
-
-async fn mark_archive_backfill_completed_for_accounts_tx(
-    tx: &mut SqliteConnection,
-    account_ids: &[i64],
-) -> Result<()> {
-    if account_ids.is_empty() {
-        return Ok(());
-    }
-    for account_chunk in account_ids.chunks(BACKFILL_ACCOUNT_BIND_BATCH_SIZE) {
-        let mut update = QueryBuilder::<Sqlite>::new(
-            "UPDATE pool_upstream_accounts SET last_activity_archive_backfill_completed = 1 WHERE id IN (",
-        );
-        {
-            let mut separated = update.separated(", ");
-            for account_id in account_chunk {
-                separated.push_bind(account_id);
-            }
-        }
-        update.push(")");
-        update.build().execute(&mut *tx).await?;
-    }
-    Ok(())
-}
-
-async fn count_upstream_accounts_missing_last_activity(pool: &Pool<Sqlite>) -> Result<u64> {
-    Ok(sqlx::query_scalar::<_, i64>(
-        r#"
-            SELECT COUNT(*)
-            FROM pool_upstream_accounts
-            WHERE last_activity_at IS NULL
-              AND last_activity_archive_backfill_completed = 0
-            "#,
-    )
-    .fetch_one(pool)
-    .await?
-    .max(0) as u64)
-}
-
-async fn count_upstream_accounts_missing_live_last_activity(pool: &Pool<Sqlite>) -> Result<u64> {
-    Ok(sqlx::query_scalar::<_, i64>(
-        r#"
-            SELECT COUNT(*)
-            FROM pool_upstream_accounts
-            WHERE last_activity_at IS NULL
-              AND last_activity_live_backfill_completed = 0
-            "#,
-    )
-    .fetch_one(pool)
-    .await?
-    .max(0) as u64)
-}
-
-async fn ensure_sqlite_file_initialized(path: &Path) -> Result<()> {
-    let database_url = format!("sqlite://{}", path.to_string_lossy());
-    let connect_opts = build_sqlite_connect_options(
-        &database_url,
-        Duration::from_secs(DEFAULT_SQLITE_BUSY_TIMEOUT_SECS),
-    )?;
-    let connection = SqliteConnection::connect_with(&connect_opts)
-        .await
-        .with_context(|| format!("failed to initialize sqlite file {}", path.display()))?;
-    connection.close().await?;
-    Ok(())
-}
-
-async fn archive_rows_into_month_batch(
-    pool: &Pool<Sqlite>,
-    config: &AppConfig,
-    spec: ArchiveTableSpec,
-    month_key: &str,
-    ids: &[i64],
-) -> Result<ArchiveBatchOutcome> {
-    if ids.is_empty() {
-        bail!("archive batch requires at least one row id");
-    }
-
-    let final_path = archive_batch_file_path(config, spec.dataset, month_key)?;
-    if let Some(parent) = final_path.parent() {
-        fs::create_dir_all(parent)
-            .with_context(|| format!("failed to create archive directory: {}", parent.display()))?;
-    }
-
-    let suffix = retention_temp_suffix();
-    let work_path = PathBuf::from(format!("{}.{}.sqlite", final_path.display(), suffix));
-    let temp_gzip_path = PathBuf::from(format!("{}.{}.tmp", final_path.display(), suffix));
-
-    if work_path.exists() {
-        let _ = fs::remove_file(&work_path);
-    }
-    if temp_gzip_path.exists() {
-        let _ = fs::remove_file(&temp_gzip_path);
-    }
-
-    if final_path.exists() {
-        inflate_gzip_sqlite_file(&final_path, &work_path)?;
-    }
-    if !work_path.exists() {
-        ensure_sqlite_file_initialized(&work_path).await?;
-    }
-
-    let row_count = async {
-        let mut conn = pool.acquire().await?;
-        sqlx::query("ATTACH DATABASE ?1 AS archive_db")
-            .bind(work_path.to_string_lossy().to_string())
-            .execute(&mut *conn)
-            .await
-            .with_context(|| {
-                format!("failed to attach archive database {}", work_path.display())
-            })?;
-        sqlx::query(spec.create_sql)
-            .execute(&mut *conn)
-            .await
-            .with_context(|| format!("failed to ensure archive schema for {}", spec.dataset))?;
-        if spec.dataset == "codex_invocations" {
-            ensure_codex_invocations_archive_schema(&mut conn).await?;
-        } else if spec.dataset == "pool_upstream_request_attempts" {
-            ensure_pool_upstream_request_attempts_archive_schema(&mut conn).await?;
-        }
-
-        let upstream_last_activity = if spec.dataset == "codex_invocations" {
-            let mut rows = Vec::new();
-            for chunk in ids.chunks(BACKFILL_ACCOUNT_BIND_BATCH_SIZE) {
-                let mut query = QueryBuilder::<Sqlite>::new(
-                    "SELECT account_id, MAX(occurred_at) AS last_activity_at FROM (SELECT CASE WHEN json_valid(payload) THEN CAST(json_extract(payload, '$.upstreamAccountId') AS INTEGER) END AS account_id, occurred_at FROM main.codex_invocations WHERE id IN (",
-                );
-                {
-                    let mut separated = query.separated(", ");
-                    for id in chunk {
-                        separated.push_bind(id);
-                    }
-                }
-                query.push(")) WHERE account_id IS NOT NULL GROUP BY account_id");
-                rows.extend(
-                    query
-                        .build_query_as::<ArchivedAccountLastActivityRow>()
-                        .fetch_all(&mut *conn)
-                        .await?,
-                );
-            }
-            dedupe_archive_upstream_last_activity(
-                rows.into_iter()
-                    .map(|row| (row.account_id, row.last_activity_at)),
-            )
-        } else {
-            Vec::new()
-        };
-
-        let mut insert = QueryBuilder::<Sqlite>::new(format!(
-            "INSERT OR IGNORE INTO archive_db.{} ({}) SELECT {} FROM main.{} WHERE id IN (",
-            spec.dataset, spec.columns, spec.columns, spec.dataset
-        ));
-        {
-            let mut separated = insert.separated(", ");
-            for id in ids {
-                separated.push_bind(id);
-            }
-        }
-        insert.push(")");
-        insert.build().execute(&mut *conn).await.with_context(|| {
-            format!(
-                "failed to copy rows into archive batch for {}",
-                spec.dataset
-            )
-        })?;
-
-        let count_query = format!("SELECT COUNT(*) FROM archive_db.{}", spec.dataset);
-        let row_count = sqlx::query_scalar::<_, i64>(&count_query)
-            .fetch_one(&mut *conn)
-            .await
-            .with_context(|| format!("failed to count archive rows for {}", spec.dataset))?;
-        sqlx::query("DETACH DATABASE archive_db")
-            .execute(&mut *conn)
-            .await
-            .context("failed to detach archive database")?;
-        Ok::<(i64, Vec<(i64, String)>), anyhow::Error>((row_count, upstream_last_activity))
-    }
-    .await;
-
-    let (result, upstream_last_activity) = match row_count {
-        Ok(values) => values,
-        Err(err) => {
-            let _ = fs::remove_file(&work_path);
-            let _ = fs::remove_file(&temp_gzip_path);
-            return Err(err);
-        }
-    };
-
-    if let Err(err) = deflate_sqlite_file_to_gzip(&work_path, &temp_gzip_path) {
-        let _ = fs::remove_file(&work_path);
-        let _ = fs::remove_file(&temp_gzip_path);
-        return Err(err);
-    }
-    if let Err(err) = fs::rename(&temp_gzip_path, &final_path).with_context(|| {
-        format!(
-            "failed to move archive batch into place: {} -> {}",
-            temp_gzip_path.display(),
-            final_path.display()
-        )
-    }) {
-        let _ = fs::remove_file(&work_path);
-        let _ = fs::remove_file(&temp_gzip_path);
-        return Err(err);
-    }
-    let _ = fs::remove_file(&work_path);
-
-    let sha256 = sha256_hex_file(&final_path)?;
-    Ok(ArchiveBatchOutcome {
-        dataset: spec.dataset,
-        month_key: month_key.to_string(),
-        day_key: None,
-        part_key: None,
-        file_path: final_path.to_string_lossy().to_string(),
-        sha256,
-        row_count: result,
-        upstream_last_activity,
-        coverage_start_at: None,
-        coverage_end_at: None,
-        archive_expires_at: None,
-        layout: ARCHIVE_LAYOUT_LEGACY_MONTH,
-        codec: ARCHIVE_FILE_CODEC_GZIP,
-        writer_version: ARCHIVE_WRITER_VERSION_LEGACY_MONTH_V1,
-        cleanup_state: ARCHIVE_CLEANUP_STATE_ACTIVE,
-        superseded_by: None,
-    })
-}
-
-async fn archive_rows_into_segment_batch(
-    pool: &Pool<Sqlite>,
-    config: &AppConfig,
-    spec: ArchiveTableSpec,
-    day_key: &str,
-    ids: &[i64],
-) -> Result<ArchiveBatchOutcome> {
-    if ids.is_empty() {
-        bail!("archive segment requires at least one row id");
-    }
-    if spec.dataset != "codex_invocations" {
-        bail!("archive segment writer only supports codex_invocations");
-    }
-    let month_key = archive_month_key_from_day_key(day_key)?;
-    let part_key = next_archive_segment_part_key(pool, spec.dataset, day_key).await?;
-    let final_path = archive_segment_file_path(
-        config,
-        spec.dataset,
-        day_key,
-        &part_key,
-        config.invocation_archive_codec,
-    )?;
-    if let Some(parent) = final_path.parent() {
-        fs::create_dir_all(parent)
-            .with_context(|| format!("failed to create archive directory: {}", parent.display()))?;
-    }
-
-    let suffix = retention_temp_suffix();
-    let work_path = PathBuf::from(format!(
-        "{}.{}.partial.sqlite",
-        final_path.display(),
-        suffix
-    ));
-    let temp_gzip_path = PathBuf::from(format!("{}.{}.tmp", final_path.display(), suffix));
-    let _temp_cleanup = TempSqliteCleanup(work_path.clone());
-    let _gzip_cleanup = TempSqliteCleanup(temp_gzip_path.clone());
-    ensure_sqlite_file_initialized(&work_path).await?;
-
-    let row_count = async {
-        let mut conn = pool.acquire().await?;
-        sqlx::query("ATTACH DATABASE ?1 AS archive_db")
-            .bind(work_path.to_string_lossy().to_string())
-            .execute(&mut *conn)
-            .await
-            .with_context(|| {
-                format!("failed to attach archive database {}", work_path.display())
-            })?;
-        sqlx::query(spec.create_sql)
-            .execute(&mut *conn)
-            .await
-            .with_context(|| format!("failed to ensure archive schema for {}", spec.dataset))?;
-        ensure_codex_invocations_archive_schema(&mut conn).await?;
-
-        let mut upstream_last_activity = Vec::new();
-        for chunk in ids.chunks(BACKFILL_ACCOUNT_BIND_BATCH_SIZE) {
-            let mut query = QueryBuilder::<Sqlite>::new(
-                "SELECT account_id, MAX(occurred_at) AS last_activity_at FROM (SELECT CASE WHEN json_valid(payload) THEN CAST(json_extract(payload, '$.upstreamAccountId') AS INTEGER) END AS account_id, occurred_at FROM main.codex_invocations WHERE id IN (",
-            );
-            {
-                let mut separated = query.separated(", ");
-                for id in chunk {
-                    separated.push_bind(id);
-                }
-            }
-            query.push(")) WHERE account_id IS NOT NULL GROUP BY account_id");
-            upstream_last_activity.extend(
-                query
-                    .build_query_as::<ArchivedAccountLastActivityRow>()
-                    .fetch_all(&mut *conn)
-                    .await?,
-            );
-        }
-        let upstream_last_activity = dedupe_archive_upstream_last_activity(
-            upstream_last_activity
-                .into_iter()
-                .map(|row| (row.account_id, row.last_activity_at)),
-        );
-
-        let mut insert = QueryBuilder::<Sqlite>::new(format!(
-            "INSERT OR IGNORE INTO archive_db.{} ({}) SELECT {} FROM main.{} WHERE id IN (",
-            spec.dataset, spec.columns, spec.columns, spec.dataset
-        ));
-        {
-            let mut separated = insert.separated(", ");
-            for id in ids {
-                separated.push_bind(id);
-            }
-        }
-        insert.push(")");
-        insert.build().execute(&mut *conn).await.with_context(|| {
-            format!(
-                "failed to copy rows into archive segment for {}",
-                spec.dataset
-            )
-        })?;
-
-        let count_query = format!("SELECT COUNT(*) FROM archive_db.{}", spec.dataset);
-        let row_count = sqlx::query_scalar::<_, i64>(&count_query)
-            .fetch_one(&mut *conn)
-            .await
-            .with_context(|| format!("failed to count archive rows for {}", spec.dataset))?;
-        sqlx::query("DETACH DATABASE archive_db")
-            .execute(&mut *conn)
-            .await
-            .context("failed to detach archive database")?;
-        Ok::<(i64, Vec<(i64, String)>), anyhow::Error>((row_count, upstream_last_activity))
-    }
-    .await?;
-
-    deflate_sqlite_file_to_gzip(&work_path, &temp_gzip_path)?;
-    fs::rename(&temp_gzip_path, &final_path).with_context(|| {
-        format!(
-            "failed to move archive segment into place: {} -> {}",
-            temp_gzip_path.display(),
-            final_path.display()
-        )
-    })?;
-
-    let sha256 = sha256_hex_file(&final_path)?;
-    Ok(ArchiveBatchOutcome {
-        dataset: spec.dataset,
-        month_key,
-        day_key: Some(day_key.to_string()),
-        part_key: Some(part_key),
-        file_path: final_path.to_string_lossy().to_string(),
-        sha256,
-        row_count: row_count.0,
-        upstream_last_activity: row_count.1,
-        coverage_start_at: None,
-        coverage_end_at: None,
-        archive_expires_at: None,
-        layout: ARCHIVE_LAYOUT_SEGMENT_V1,
-        codec: config.invocation_archive_codec.as_str(),
-        writer_version: ARCHIVE_WRITER_VERSION_SEGMENT_V1,
-        cleanup_state: ARCHIVE_CLEANUP_STATE_ACTIVE,
-        superseded_by: None,
-    })
-}
-
-async fn upsert_archived_upstream_last_activity(
-    tx: &mut sqlx::SqliteConnection,
-    values: &[(i64, String)],
-) -> Result<()> {
-    let deduped_values = dedupe_archive_upstream_last_activity(values.iter().cloned());
-    for (account_id, occurred_at) in &deduped_values {
-        sqlx::query(
-            r#"
-            UPDATE pool_upstream_accounts
-            SET last_activity_at = CASE
-                    WHEN last_activity_at IS NULL OR last_activity_at < ?1 THEN ?1
-                    ELSE last_activity_at
-                END
-            WHERE id = ?2
-            "#,
-        )
-        .bind(occurred_at)
-        .bind(account_id)
-        .execute(&mut *tx)
-        .await?;
-    }
-    Ok(())
-}
-
-async fn upsert_archive_batch_manifest(
-    tx: &mut sqlx::SqliteConnection,
-    batch: &ArchiveBatchOutcome,
-) -> Result<()> {
-    sqlx::query(
-        r#"
-        INSERT INTO archive_batches (
-            dataset,
-            month_key,
-            day_key,
-            part_key,
-            file_path,
-            sha256,
-            row_count,
-            status,
-            layout,
-            codec,
-            writer_version,
-            cleanup_state,
-            superseded_by,
-            coverage_start_at,
-            coverage_end_at,
-            archive_expires_at,
-            created_at
-        )
-        VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8, ?9, ?10, ?11, ?12, ?13, ?14, ?15, ?16, datetime('now'))
-        ON CONFLICT(dataset, month_key, file_path) DO UPDATE SET
-            day_key = excluded.day_key,
-            part_key = excluded.part_key,
-            sha256 = excluded.sha256,
-            row_count = excluded.row_count,
-            status = excluded.status,
-            layout = excluded.layout,
-            codec = excluded.codec,
-            writer_version = excluded.writer_version,
-            cleanup_state = excluded.cleanup_state,
-            superseded_by = excluded.superseded_by,
-            coverage_start_at = excluded.coverage_start_at,
-            coverage_end_at = excluded.coverage_end_at,
-            archive_expires_at = excluded.archive_expires_at,
-            created_at = datetime('now')
-        "#,
-    )
-    .bind(batch.dataset)
-    .bind(&batch.month_key)
-    .bind(batch.day_key.as_deref())
-    .bind(batch.part_key.as_deref())
-    .bind(&batch.file_path)
-    .bind(&batch.sha256)
-    .bind(batch.row_count)
-    .bind(ARCHIVE_STATUS_COMPLETED)
-    .bind(batch.layout)
-    .bind(batch.codec)
-    .bind(batch.writer_version)
-    .bind(batch.cleanup_state)
-    .bind(batch.superseded_by)
-    .bind(batch.coverage_start_at.as_deref())
-    .bind(batch.coverage_end_at.as_deref())
-    .bind(batch.archive_expires_at.as_deref())
-    .execute(&mut *tx)
-    .await?;
-    let deduped_upstream_last_activity =
-        dedupe_archive_upstream_last_activity(batch.upstream_last_activity.iter().cloned());
-    let archive_batch_id = sqlx::query_scalar::<_, i64>(
-        r#"
-        SELECT id
-        FROM archive_batches
-        WHERE dataset = ?1
-          AND month_key = ?2
-          AND file_path = ?3
-        LIMIT 1
-        "#,
-    )
-    .bind(batch.dataset)
-    .bind(&batch.month_key)
-    .bind(&batch.file_path)
-    .fetch_one(&mut *tx)
-    .await?;
-    if batch.dataset == "codex_invocations" {
-        write_archive_batch_upstream_activity(
-            tx,
-            archive_batch_id,
-            &deduped_upstream_last_activity,
-        )
-        .await?;
-    }
-    if batch.dataset == "codex_invocations" && !deduped_upstream_last_activity.is_empty() {
-        upsert_archived_upstream_last_activity(tx, &deduped_upstream_last_activity).await?;
-    }
-    Ok(())
-}
-
-async fn write_archive_batch_upstream_activity(
-    tx: &mut sqlx::SqliteConnection,
-    archive_batch_id: i64,
-    values: &[(i64, String)],
-) -> Result<()> {
-    let deduped_values = dedupe_archive_upstream_last_activity(values.iter().cloned());
-    sqlx::query("DELETE FROM archive_batch_upstream_activity WHERE archive_batch_id = ?1")
-        .bind(archive_batch_id)
-        .execute(&mut *tx)
-        .await?;
-    for chunk in deduped_values.chunks(BACKFILL_ACCOUNT_BIND_BATCH_SIZE) {
-        let mut insert = QueryBuilder::<Sqlite>::new(
-            "INSERT INTO archive_batch_upstream_activity (archive_batch_id, account_id, last_activity_at) ",
-        );
-        insert.push_values(chunk, |mut row, (account_id, last_activity_at)| {
-            row.push_bind(archive_batch_id)
-                .push_bind(account_id)
-                .push_bind(last_activity_at);
-        });
-        insert.push(
-            " ON CONFLICT(archive_batch_id, account_id) DO UPDATE SET last_activity_at = CASE \
-             WHEN excluded.last_activity_at > last_activity_at THEN excluded.last_activity_at \
-             ELSE last_activity_at END",
-        );
-        insert.build().execute(&mut *tx).await?;
-    }
-    sqlx::query(
-        "UPDATE archive_batches SET upstream_activity_manifest_refreshed_at = datetime('now') WHERE id = ?1",
-    )
-    .bind(archive_batch_id)
-    .execute(&mut *tx)
-    .await?;
-    Ok(())
-}
-
-async fn mark_retention_archived_hourly_rollup_targets_tx(
-    tx: &mut SqliteConnection,
-    dataset: &str,
-    invocation_rows: &[InvocationHourlySourceRecord],
-    forward_proxy_rows: &[ForwardProxyAttemptHourlySourceRecord],
-) -> Result<()> {
-    match dataset {
-        "codex_invocations" => {
-            mark_invocation_hourly_rollup_buckets_materialized_tx(tx, invocation_rows).await?;
-        }
-        "forward_proxy_attempts" => {
-            mark_forward_proxy_hourly_rollup_buckets_materialized_tx(tx, forward_proxy_rows)
-                .await?;
-        }
-        _ => {}
-    }
-    Ok(())
-}
-
-async fn mark_archive_batch_historical_rollups_materialized_tx(
-    tx: &mut SqliteConnection,
-    dataset: &str,
-    file_path: &str,
-) -> Result<()> {
-    sqlx::query(
-        r#"
-        UPDATE archive_batches
-        SET historical_rollups_materialized_at = datetime('now')
-        WHERE dataset = ?1
-          AND file_path = ?2
-        "#,
-    )
-    .bind(dataset)
-    .bind(file_path)
-    .execute(&mut *tx)
-    .await?;
-    Ok(())
-}
-
-async fn update_archive_batch_coverage_bounds_tx(
-    tx: &mut SqliteConnection,
-    archive_batch_id: i64,
-    coverage_start_at: Option<&str>,
-    coverage_end_at: Option<&str>,
-) -> Result<()> {
-    sqlx::query(
-        r#"
-        UPDATE archive_batches
-        SET coverage_start_at = COALESCE(coverage_start_at, ?2),
-            coverage_end_at = COALESCE(coverage_end_at, ?3)
-        WHERE id = ?1
-        "#,
-    )
-    .bind(archive_batch_id)
-    .bind(coverage_start_at)
-    .bind(coverage_end_at)
-    .execute(&mut *tx)
-    .await?;
-    Ok(())
-}
-
-async fn mark_hourly_rollup_bucket_materialized_tx(
-    tx: &mut SqliteConnection,
-    target: &str,
-    bucket_start_epoch: i64,
-    source: &str,
-) -> Result<()> {
-    sqlx::query(
-        r#"
-        INSERT INTO hourly_rollup_materialized_buckets (
-            target,
-            bucket_start_epoch,
-            source,
-            materialized_at
-        )
-        VALUES (?1, ?2, ?3, datetime('now'))
-        ON CONFLICT(target, bucket_start_epoch, source) DO UPDATE SET
-            materialized_at = datetime('now')
-        "#,
-    )
-    .bind(target)
-    .bind(bucket_start_epoch)
-    .bind(source)
-    .execute(&mut *tx)
-    .await?;
-    Ok(())
-}
-
-async fn mark_invocation_hourly_rollup_buckets_materialized_tx(
-    tx: &mut SqliteConnection,
-    rows: &[InvocationHourlySourceRecord],
-) -> Result<()> {
-    let mut overall_targets = HashSet::new();
-    let mut sticky_targets = HashSet::new();
-    for row in rows {
-        let bucket_start_epoch = invocation_bucket_start_epoch(&row.occurred_at)?;
-        overall_targets.insert((bucket_start_epoch, row.source.clone()));
-        sticky_targets.insert(bucket_start_epoch);
-    }
-
-    let live_targets = load_live_invocation_bucket_targets_tx(tx, &overall_targets).await?;
-    let live_proxy_buckets = live_targets
-        .iter()
-        .filter_map(|(bucket_start_epoch, source)| {
-            (source == SOURCE_PROXY).then_some(*bucket_start_epoch)
-        })
-        .collect::<HashSet<_>>();
-
-    for (bucket_start_epoch, source) in overall_targets {
-        if live_targets.contains(&(bucket_start_epoch, source.clone())) {
-            continue;
-        }
-        for target in [
-            HOURLY_ROLLUP_TARGET_INVOCATIONS,
-            HOURLY_ROLLUP_TARGET_INVOCATION_FAILURES,
-            HOURLY_ROLLUP_TARGET_PROMPT_CACHE,
-            HOURLY_ROLLUP_TARGET_PROMPT_CACHE_UPSTREAM_ACCOUNTS,
-        ] {
-            mark_hourly_rollup_bucket_materialized_tx(tx, target, bucket_start_epoch, &source)
-                .await?;
-        }
-        if source == SOURCE_PROXY && !live_proxy_buckets.contains(&bucket_start_epoch) {
-            mark_hourly_rollup_bucket_materialized_tx(
-                tx,
-                HOURLY_ROLLUP_TARGET_PROXY_PERF,
-                bucket_start_epoch,
-                SOURCE_PROXY,
-            )
-            .await?;
-        }
-    }
-
-    for bucket_start_epoch in sticky_targets {
-        if live_proxy_buckets.contains(&bucket_start_epoch) {
-            continue;
-        }
-        mark_hourly_rollup_bucket_materialized_tx(
-            tx,
-            HOURLY_ROLLUP_TARGET_STICKY_KEYS,
-            bucket_start_epoch,
-            HOURLY_ROLLUP_MATERIALIZED_SOURCE_NONE,
-        )
-        .await?;
-    }
-
-    Ok(())
-}
-
-async fn load_live_invocation_bucket_targets_tx(
-    tx: &mut SqliteConnection,
-    bucket_targets: &HashSet<(i64, String)>,
-) -> Result<HashSet<(i64, String)>> {
-    if bucket_targets.is_empty() {
-        return Ok(HashSet::new());
-    }
-
-    let min_bucket_epoch = bucket_targets
-        .iter()
-        .map(|(bucket_start_epoch, _)| *bucket_start_epoch)
-        .min()
-        .ok_or_else(|| anyhow!("missing minimum invocation bucket epoch"))?;
-    let max_bucket_epoch = bucket_targets
-        .iter()
-        .map(|(bucket_start_epoch, _)| *bucket_start_epoch)
-        .max()
-        .ok_or_else(|| anyhow!("missing maximum invocation bucket epoch"))?;
-    let min_bucket_start = Utc
-        .timestamp_opt(min_bucket_epoch, 0)
-        .single()
-        .ok_or_else(|| anyhow!("invalid minimum invocation bucket epoch"))?;
-    let max_bucket_end = Utc
-        .timestamp_opt(max_bucket_epoch + 3_600, 0)
-        .single()
-        .ok_or_else(|| anyhow!("invalid maximum invocation bucket epoch"))?;
-
-    let rows = sqlx::query_as::<_, InvocationBucketPresenceRow>(
-        r#"
-        SELECT occurred_at, source
-        FROM codex_invocations
-        WHERE occurred_at >= ?1
-          AND occurred_at < ?2
-        ORDER BY id ASC
-        "#,
-    )
-    .bind(db_occurred_at_lower_bound(min_bucket_start))
-    .bind(db_occurred_at_lower_bound(max_bucket_end))
-    .fetch_all(&mut *tx)
-    .await?;
-
-    let mut live_targets = HashSet::new();
-    for row in rows {
-        let key = (invocation_bucket_start_epoch(&row.occurred_at)?, row.source);
-        if bucket_targets.contains(&key) {
-            live_targets.insert(key);
-        }
-    }
-    Ok(live_targets)
-}
-
-async fn mark_forward_proxy_hourly_rollup_buckets_materialized_tx(
-    tx: &mut SqliteConnection,
-    rows: &[ForwardProxyAttemptHourlySourceRecord],
-) -> Result<()> {
-    let mut buckets = HashSet::new();
-    for row in rows {
-        buckets.insert(forward_proxy_attempt_bucket_start_epoch(&row.occurred_at)?);
-    }
-    for bucket_start_epoch in buckets {
-        mark_hourly_rollup_bucket_materialized_tx(
-            tx,
-            HOURLY_ROLLUP_TARGET_FORWARD_PROXY_ATTEMPTS,
-            bucket_start_epoch,
-            HOURLY_ROLLUP_MATERIALIZED_SOURCE_NONE,
-        )
-        .await?;
-    }
-    Ok(())
-}
-
-async fn upsert_invocation_rollups(
-    tx: &mut sqlx::SqliteConnection,
-    candidates: &[InvocationArchiveCandidate],
-) -> Result<()> {
-    let mut rollups: BTreeMap<(String, String), InvocationRollupDelta> = BTreeMap::new();
-    for candidate in candidates {
-        let stats_date = shanghai_day_key_from_local_naive(&candidate.occurred_at)?;
-        let key = (stats_date, candidate.source.clone());
-        let entry = rollups.entry(key).or_default();
-        entry.total_count += 1;
-        if matches!(candidate.status.as_deref(), Some("success")) {
-            entry.success_count += 1;
-        } else if candidate
-            .status
-            .as_deref()
-            .is_some_and(|status| status != "success")
-        {
-            entry.failure_count += 1;
-        }
-        entry.total_tokens += candidate.total_tokens.unwrap_or_default();
-        entry.total_cost += candidate.cost.unwrap_or_default();
-    }
-
-    for ((stats_date, source), delta) in rollups {
-        sqlx::query(
-            r#"
-            INSERT INTO invocation_rollup_daily (
-                stats_date,
-                source,
-                total_count,
-                success_count,
-                failure_count,
-                total_tokens,
-                total_cost,
-                created_at
-            )
-            VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, datetime('now'))
-            ON CONFLICT(stats_date, source) DO UPDATE SET
-                total_count = invocation_rollup_daily.total_count + excluded.total_count,
-                success_count = invocation_rollup_daily.success_count + excluded.success_count,
-                failure_count = invocation_rollup_daily.failure_count + excluded.failure_count,
-                total_tokens = invocation_rollup_daily.total_tokens + excluded.total_tokens,
-                total_cost = invocation_rollup_daily.total_cost + excluded.total_cost
-            "#,
-        )
-        .bind(&stats_date)
-        .bind(&source)
-        .bind(delta.total_count)
-        .bind(delta.success_count)
-        .bind(delta.failure_count)
-        .bind(delta.total_tokens)
-        .bind(delta.total_cost)
-        .execute(&mut *tx)
-        .await?;
-    }
-
-    Ok(())
-}
-
-#[path = "archive_hourly_rollup_support.rs"]
-mod archive_hourly_rollup_support;
-pub(crate) use archive_hourly_rollup_support::*;
-
-async fn upsert_invocation_hourly_rollups_tx(
-    tx: &mut SqliteConnection,
-    rows: &[InvocationHourlySourceRecord],
-    targets: &[&str],
-) -> Result<()> {
-    if rows.is_empty() {
-        return Ok(());
-    }
-    let upsert_overall = targets.contains(&HOURLY_ROLLUP_TARGET_INVOCATIONS);
-    let upsert_failures = targets.contains(&HOURLY_ROLLUP_TARGET_INVOCATION_FAILURES);
-    let upsert_perf = targets.contains(&HOURLY_ROLLUP_TARGET_PROXY_PERF);
-    let upsert_prompt_cache = targets.contains(&HOURLY_ROLLUP_TARGET_PROMPT_CACHE);
-    let upsert_prompt_cache_upstream_accounts =
-        targets.contains(&HOURLY_ROLLUP_TARGET_PROMPT_CACHE_UPSTREAM_ACCOUNTS);
-    let upsert_sticky_keys = targets.contains(&HOURLY_ROLLUP_TARGET_STICKY_KEYS);
-
-    let mut overall: BTreeMap<(i64, String), InvocationHourlyRollupDelta> = BTreeMap::new();
-    let mut failures: BTreeMap<(i64, String, String, i64, String), i64> = BTreeMap::new();
-    let mut perf: BTreeMap<(i64, String), ProxyPerfStageHourlyDelta> = BTreeMap::new();
-    let mut prompt_cache: BTreeMap<(i64, String, String), KeyedConversationHourlyDelta> =
-        BTreeMap::new();
-    let mut prompt_cache_upstream_accounts: BTreeMap<
-        (i64, String, String, String, Option<i64>, Option<String>),
-        KeyedConversationHourlyDelta,
-    > = BTreeMap::new();
-    let mut sticky_keys: BTreeMap<(i64, i64, String), KeyedConversationHourlyDelta> =
-        BTreeMap::new();
-
-    for row in rows {
-        let bucket_start_epoch = invocation_bucket_start_epoch(&row.occurred_at)?;
-        if upsert_overall {
-            accumulate_invocation_hourly_overall_rollups(&mut overall, std::slice::from_ref(row))?;
-        }
-
-        if upsert_failures {
-            let classification = resolve_failure_classification(
-                row.status.as_deref(),
-                row.error_message.as_deref(),
-                row.failure_kind.as_deref(),
-                row.failure_class.as_deref(),
-                row.is_actionable,
-            );
-            if invocation_status_counts_toward_terminal_totals(row.status.as_deref())
-                && classification.failure_class != FailureClass::None
-            {
-                let error_category =
-                    categorize_error(row.error_message.as_deref().unwrap_or_default());
-                *failures
-                    .entry((
-                        bucket_start_epoch,
-                        row.source.clone(),
-                        classification.failure_class.as_str().to_string(),
-                        classification.is_actionable as i64,
-                        error_category,
-                    ))
-                    .or_default() += 1;
-            }
-        }
-
-        if upsert_perf && row.source == SOURCE_PROXY {
-            record_proxy_perf_stage_sample(
-                &mut perf,
-                bucket_start_epoch,
-                PERF_STAGE_TOTAL,
-                row.t_total_ms,
-            );
-            record_proxy_perf_stage_sample(
-                &mut perf,
-                bucket_start_epoch,
-                PERF_STAGE_REQUEST_READ,
-                row.t_req_read_ms,
-            );
-            record_proxy_perf_stage_sample(
-                &mut perf,
-                bucket_start_epoch,
-                PERF_STAGE_REQUEST_PARSE,
-                row.t_req_parse_ms,
-            );
-            record_proxy_perf_stage_sample(
-                &mut perf,
-                bucket_start_epoch,
-                PERF_STAGE_UPSTREAM_CONNECT,
-                row.t_upstream_connect_ms,
-            );
-            record_proxy_perf_stage_sample(
-                &mut perf,
-                bucket_start_epoch,
-                PERF_STAGE_UPSTREAM_FIRST_BYTE,
-                row.t_upstream_ttfb_ms,
-            );
-            record_proxy_perf_stage_sample(
-                &mut perf,
-                bucket_start_epoch,
-                PERF_STAGE_UPSTREAM_STREAM,
-                row.t_upstream_stream_ms,
-            );
-            record_proxy_perf_stage_sample(
-                &mut perf,
-                bucket_start_epoch,
-                PERF_STAGE_RESPONSE_PARSE,
-                row.t_resp_parse_ms,
-            );
-            record_proxy_perf_stage_sample(
-                &mut perf,
-                bucket_start_epoch,
-                PERF_STAGE_PERSISTENCE,
-                row.t_persist_ms,
-            );
-        }
-
-        if (upsert_prompt_cache || upsert_prompt_cache_upstream_accounts)
-            && let Some(prompt_cache_key) = prompt_cache_key_from_payload(row.payload.as_deref())
-        {
-            if upsert_prompt_cache {
-                let entry = keyed_conversation_delta(
-                    &mut prompt_cache,
-                    bucket_start_epoch,
-                    &row.source,
-                    &prompt_cache_key,
-                    &row.occurred_at,
-                );
-                entry.request_count += 1;
-                if row.status.as_deref() == Some("success") {
-                    entry.success_count += 1;
-                } else {
-                    entry.failure_count += 1;
-                }
-                entry.total_tokens += row.total_tokens.unwrap_or_default();
-                entry.total_cost += row.cost.unwrap_or_default();
-            }
-
-            if upsert_prompt_cache_upstream_accounts {
-                let upstream_account_id = upstream_account_id_from_payload(row.payload.as_deref());
-                let upstream_account_name =
-                    upstream_account_name_from_payload(row.payload.as_deref());
-                let rollup_key = prompt_cache_upstream_account_rollup_key(
-                    upstream_account_id,
-                    upstream_account_name.as_deref(),
-                );
-                let entry = prompt_cache_upstream_accounts
-                    .entry((
-                        bucket_start_epoch,
-                        row.source.clone(),
-                        prompt_cache_key,
-                        rollup_key,
-                        upstream_account_id,
-                        upstream_account_name.clone(),
-                    ))
-                    .or_insert_with(|| KeyedConversationHourlyDelta {
-                        first_seen_at: row.occurred_at.clone(),
-                        last_seen_at: row.occurred_at.clone(),
-                        ..KeyedConversationHourlyDelta::default()
-                    });
-                if row.occurred_at < entry.first_seen_at {
-                    entry.first_seen_at = row.occurred_at.clone();
-                }
-                if row.occurred_at > entry.last_seen_at {
-                    entry.last_seen_at = row.occurred_at.clone();
-                }
-                entry.request_count += 1;
-                if row.status.as_deref() == Some("success") {
-                    entry.success_count += 1;
-                } else {
-                    entry.failure_count += 1;
-                }
-                entry.total_tokens += row.total_tokens.unwrap_or_default();
-                entry.total_cost += row.cost.unwrap_or_default();
-            }
-        }
-
-        if upsert_sticky_keys
-            && let (Some(upstream_account_id), Some(sticky_key)) = (
-                upstream_account_id_from_payload(row.payload.as_deref()),
-                sticky_key_from_payload(row.payload.as_deref()),
-            )
-        {
-            let entry = sticky_keys
-                .entry((bucket_start_epoch, upstream_account_id, sticky_key))
-                .or_insert_with(|| KeyedConversationHourlyDelta {
-                    first_seen_at: row.occurred_at.clone(),
-                    last_seen_at: row.occurred_at.clone(),
-                    ..KeyedConversationHourlyDelta::default()
-                });
-            if row.occurred_at < entry.first_seen_at {
-                entry.first_seen_at = row.occurred_at.clone();
-            }
-            if row.occurred_at > entry.last_seen_at {
-                entry.last_seen_at = row.occurred_at.clone();
-            }
-            entry.request_count += 1;
-            if row.status.as_deref() == Some("success") {
-                entry.success_count += 1;
-            } else {
-                entry.failure_count += 1;
-            }
-            entry.total_tokens += row.total_tokens.unwrap_or_default();
-            entry.total_cost += row.cost.unwrap_or_default();
-        }
-    }
-
-    if upsert_overall {
-        #[derive(sqlx::FromRow)]
-        struct InvocationRollupHistogramRow {
-            first_byte_histogram: String,
-            first_response_byte_total_histogram: String,
-        }
-
-        for ((bucket_start_epoch, source), delta) in overall {
-            let current_histograms = sqlx::query_as::<_, InvocationRollupHistogramRow>(
-                r#"
-                SELECT
-                    first_byte_histogram,
-                    first_response_byte_total_histogram
-                FROM invocation_rollup_hourly
-                WHERE bucket_start_epoch = ?1 AND source = ?2
-                "#,
-            )
-            .bind(bucket_start_epoch)
-            .bind(&source)
-            .fetch_optional(&mut *tx)
-            .await?;
-            let mut merged_first_byte_histogram = current_histograms
-                .as_ref()
-                .map(|row| decode_approx_histogram(&row.first_byte_histogram))
-                .unwrap_or_else(empty_approx_histogram);
-            merge_approx_histogram_into(
-                &mut merged_first_byte_histogram,
-                &delta.first_byte_histogram,
-            )?;
-            let mut merged_first_response_byte_total_histogram = current_histograms
-                .as_ref()
-                .map(|row| decode_approx_histogram(&row.first_response_byte_total_histogram))
-                .unwrap_or_else(empty_approx_histogram);
-            merge_approx_histogram_into(
-                &mut merged_first_response_byte_total_histogram,
-                &delta.first_response_byte_total_histogram,
-            )?;
-            sqlx::query(
-                r#"
-                INSERT INTO invocation_rollup_hourly (
-                    bucket_start_epoch,
-                    source,
-                    total_count,
-                    success_count,
-                    failure_count,
-                    total_tokens,
-                    total_cost,
-                    first_byte_sample_count,
-                    first_byte_sum_ms,
-                    first_byte_max_ms,
-                    first_byte_histogram,
-                    first_response_byte_total_sample_count,
-                    first_response_byte_total_sum_ms,
-                    first_response_byte_total_max_ms,
-                    first_response_byte_total_histogram,
-                    updated_at
-                )
-                VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8, ?9, ?10, ?11, ?12, ?13, ?14, ?15, datetime('now'))
-                ON CONFLICT(bucket_start_epoch, source) DO UPDATE SET
-                    total_count = invocation_rollup_hourly.total_count + excluded.total_count,
-                    success_count = invocation_rollup_hourly.success_count + excluded.success_count,
-                    failure_count = invocation_rollup_hourly.failure_count + excluded.failure_count,
-                    total_tokens = invocation_rollup_hourly.total_tokens + excluded.total_tokens,
-                    total_cost = invocation_rollup_hourly.total_cost + excluded.total_cost,
-                    first_byte_sample_count = invocation_rollup_hourly.first_byte_sample_count + excluded.first_byte_sample_count,
-                    first_byte_sum_ms = invocation_rollup_hourly.first_byte_sum_ms + excluded.first_byte_sum_ms,
-                    first_byte_max_ms = MAX(invocation_rollup_hourly.first_byte_max_ms, excluded.first_byte_max_ms),
-                    first_byte_histogram = excluded.first_byte_histogram,
-                    first_response_byte_total_sample_count = invocation_rollup_hourly.first_response_byte_total_sample_count + excluded.first_response_byte_total_sample_count,
-                    first_response_byte_total_sum_ms = invocation_rollup_hourly.first_response_byte_total_sum_ms + excluded.first_response_byte_total_sum_ms,
-                    first_response_byte_total_max_ms = MAX(invocation_rollup_hourly.first_response_byte_total_max_ms, excluded.first_response_byte_total_max_ms),
-                    first_response_byte_total_histogram = excluded.first_response_byte_total_histogram,
-                    updated_at = datetime('now')
-                "#,
-            )
-            .bind(bucket_start_epoch)
-            .bind(&source)
-            .bind(delta.total_count)
-            .bind(delta.success_count)
-            .bind(delta.failure_count)
-            .bind(delta.total_tokens)
-            .bind(delta.total_cost)
-            .bind(delta.first_byte_sample_count)
-            .bind(delta.first_byte_sum_ms)
-            .bind(delta.first_byte_max_ms)
-            .bind(encode_approx_histogram(&merged_first_byte_histogram)?)
-            .bind(delta.first_response_byte_total_sample_count)
-            .bind(delta.first_response_byte_total_sum_ms)
-            .bind(delta.first_response_byte_total_max_ms)
-            .bind(encode_approx_histogram(
-                &merged_first_response_byte_total_histogram,
-            )?)
-            .execute(&mut *tx)
-            .await?;
-        }
-    }
-
-    if upsert_failures {
-        for (
-            (bucket_start_epoch, source, failure_class, is_actionable, error_category),
-            failure_count,
-        ) in failures
-        {
-            sqlx::query(
-                r#"
-                INSERT INTO invocation_failure_rollup_hourly (
-                    bucket_start_epoch,
-                    source,
-                    failure_class,
-                    is_actionable,
-                    error_category,
-                    failure_count,
-                    updated_at
-                )
-                VALUES (?1, ?2, ?3, ?4, ?5, ?6, datetime('now'))
-                ON CONFLICT(bucket_start_epoch, source, failure_class, is_actionable, error_category) DO UPDATE SET
-                    failure_count = invocation_failure_rollup_hourly.failure_count + excluded.failure_count,
-                    updated_at = datetime('now')
-                "#,
-            )
-            .bind(bucket_start_epoch)
-            .bind(&source)
-            .bind(&failure_class)
-            .bind(is_actionable)
-            .bind(&error_category)
-            .bind(failure_count)
-            .execute(&mut *tx)
-            .await?;
-        }
-    }
-
-    if upsert_perf {
-        for ((bucket_start_epoch, stage), delta) in perf {
-            let current_histogram = sqlx::query_scalar::<_, String>(
-                r#"
-                SELECT histogram
-                FROM proxy_perf_stage_hourly
-                WHERE bucket_start_epoch = ?1 AND stage = ?2
-                "#,
-            )
-            .bind(bucket_start_epoch)
-            .bind(&stage)
-            .fetch_optional(&mut *tx)
-            .await?;
-            let mut merged_histogram = current_histogram
-                .as_deref()
-                .map(decode_approx_histogram)
-                .unwrap_or_else(empty_approx_histogram);
-            merge_approx_histogram_into(&mut merged_histogram, &delta.histogram)?;
-            sqlx::query(
-                r#"
-                INSERT INTO proxy_perf_stage_hourly (
-                    bucket_start_epoch,
-                    stage,
-                    sample_count,
-                    sum_ms,
-                    max_ms,
-                    histogram,
-                    updated_at
-                )
-                VALUES (?1, ?2, ?3, ?4, ?5, ?6, datetime('now'))
-                ON CONFLICT(bucket_start_epoch, stage) DO UPDATE SET
-                    sample_count = proxy_perf_stage_hourly.sample_count + excluded.sample_count,
-                    sum_ms = proxy_perf_stage_hourly.sum_ms + excluded.sum_ms,
-                    max_ms = MAX(proxy_perf_stage_hourly.max_ms, excluded.max_ms),
-                    histogram = excluded.histogram,
-                    updated_at = datetime('now')
-                "#,
-            )
-            .bind(bucket_start_epoch)
-            .bind(&stage)
-            .bind(delta.sample_count)
-            .bind(delta.sum_ms)
-            .bind(delta.max_ms)
-            .bind(encode_approx_histogram(&merged_histogram)?)
-            .execute(&mut *tx)
-            .await?;
-        }
-    }
-
-    if upsert_prompt_cache {
-        for ((bucket_start_epoch, source, prompt_cache_key), delta) in prompt_cache {
-            sqlx::query(
-                r#"
-                INSERT INTO prompt_cache_rollup_hourly (
-                    bucket_start_epoch,
-                    source,
-                    prompt_cache_key,
-                    request_count,
-                    success_count,
-                    failure_count,
-                    total_tokens,
-                    total_cost,
-                    first_seen_at,
-                    last_seen_at,
-                    updated_at
-                )
-                VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8, ?9, ?10, datetime('now'))
-                ON CONFLICT(bucket_start_epoch, source, prompt_cache_key) DO UPDATE SET
-                    request_count = prompt_cache_rollup_hourly.request_count + excluded.request_count,
-                    success_count = prompt_cache_rollup_hourly.success_count + excluded.success_count,
-                    failure_count = prompt_cache_rollup_hourly.failure_count + excluded.failure_count,
-                    total_tokens = prompt_cache_rollup_hourly.total_tokens + excluded.total_tokens,
-                    total_cost = prompt_cache_rollup_hourly.total_cost + excluded.total_cost,
-                    first_seen_at = MIN(prompt_cache_rollup_hourly.first_seen_at, excluded.first_seen_at),
-                    last_seen_at = MAX(prompt_cache_rollup_hourly.last_seen_at, excluded.last_seen_at),
-                    updated_at = datetime('now')
-                "#,
-            )
-            .bind(bucket_start_epoch)
-            .bind(&source)
-            .bind(&prompt_cache_key)
-            .bind(delta.request_count)
-            .bind(delta.success_count)
-            .bind(delta.failure_count)
-            .bind(delta.total_tokens)
-            .bind(delta.total_cost)
-            .bind(&delta.first_seen_at)
-            .bind(&delta.last_seen_at)
-            .execute(&mut *tx)
-            .await?;
-        }
-    }
-
-    if upsert_prompt_cache_upstream_accounts {
-        for (
-            (
-                bucket_start_epoch,
-                source,
-                prompt_cache_key,
-                upstream_account_key,
-                upstream_account_id,
-                upstream_account_name,
-            ),
-            delta,
-        ) in prompt_cache_upstream_accounts
-        {
-            sqlx::query(
-                r#"
-                INSERT INTO prompt_cache_upstream_account_hourly (
-                    bucket_start_epoch,
-                    source,
-                    prompt_cache_key,
-                    upstream_account_key,
-                    upstream_account_id,
-                    upstream_account_name,
-                    request_count,
-                    success_count,
-                    failure_count,
-                    total_tokens,
-                    total_cost,
-                    first_seen_at,
-                    last_seen_at,
-                    updated_at
-                )
-                VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8, ?9, ?10, ?11, ?12, ?13, datetime('now'))
-                ON CONFLICT(bucket_start_epoch, source, prompt_cache_key, upstream_account_key) DO UPDATE SET
-                    request_count = prompt_cache_upstream_account_hourly.request_count + excluded.request_count,
-                    success_count = prompt_cache_upstream_account_hourly.success_count + excluded.success_count,
-                    failure_count = prompt_cache_upstream_account_hourly.failure_count + excluded.failure_count,
-                    total_tokens = prompt_cache_upstream_account_hourly.total_tokens + excluded.total_tokens,
-                    total_cost = prompt_cache_upstream_account_hourly.total_cost + excluded.total_cost,
-                    first_seen_at = MIN(prompt_cache_upstream_account_hourly.first_seen_at, excluded.first_seen_at),
-                    last_seen_at = MAX(prompt_cache_upstream_account_hourly.last_seen_at, excluded.last_seen_at),
-                    updated_at = datetime('now')
-                "#,
-            )
-            .bind(bucket_start_epoch)
-            .bind(&source)
-            .bind(&prompt_cache_key)
-            .bind(&upstream_account_key)
-            .bind(upstream_account_id)
-            .bind(upstream_account_name.as_deref())
-            .bind(delta.request_count)
-            .bind(delta.success_count)
-            .bind(delta.failure_count)
-            .bind(delta.total_tokens)
-            .bind(delta.total_cost)
-            .bind(&delta.first_seen_at)
-            .bind(&delta.last_seen_at)
-            .execute(&mut *tx)
-            .await?;
-        }
-    }
-
-    if upsert_sticky_keys {
-        for ((bucket_start_epoch, upstream_account_id, sticky_key), delta) in sticky_keys {
-            sqlx::query(
-                r#"
-                INSERT INTO upstream_sticky_key_hourly (
-                    bucket_start_epoch,
-                    upstream_account_id,
-                    sticky_key,
-                    request_count,
-                    success_count,
-                    failure_count,
-                    total_tokens,
-                    total_cost,
-                    first_seen_at,
-                    last_seen_at,
-                    updated_at
-                )
-                VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8, ?9, ?10, datetime('now'))
-                ON CONFLICT(bucket_start_epoch, upstream_account_id, sticky_key) DO UPDATE SET
-                    request_count = upstream_sticky_key_hourly.request_count + excluded.request_count,
-                    success_count = upstream_sticky_key_hourly.success_count + excluded.success_count,
-                    failure_count = upstream_sticky_key_hourly.failure_count + excluded.failure_count,
-                    total_tokens = upstream_sticky_key_hourly.total_tokens + excluded.total_tokens,
-                    total_cost = upstream_sticky_key_hourly.total_cost + excluded.total_cost,
-                    first_seen_at = MIN(upstream_sticky_key_hourly.first_seen_at, excluded.first_seen_at),
-                    last_seen_at = MAX(upstream_sticky_key_hourly.last_seen_at, excluded.last_seen_at),
-                    updated_at = datetime('now')
-                "#,
-            )
-            .bind(bucket_start_epoch)
-            .bind(upstream_account_id)
-            .bind(&sticky_key)
-            .bind(delta.request_count)
-            .bind(delta.success_count)
-            .bind(delta.failure_count)
-            .bind(delta.total_tokens)
-            .bind(delta.total_cost)
-            .bind(&delta.first_seen_at)
-            .bind(&delta.last_seen_at)
-            .execute(&mut *tx)
-            .await?;
-        }
-    }
-
-    Ok(())
-}
-
-fn invocation_archive_target_needs_full_payload(target: &str) -> bool {
-    matches!(
-        target,
-        HOURLY_ROLLUP_TARGET_PROMPT_CACHE
-            | HOURLY_ROLLUP_TARGET_PROMPT_CACHE_UPSTREAM_ACCOUNTS
-            | HOURLY_ROLLUP_TARGET_STICKY_KEYS
-    )
-}
-
-async fn upsert_forward_proxy_attempt_hourly_rollups_tx(
-    tx: &mut SqliteConnection,
-    rows: &[ForwardProxyAttemptHourlySourceRecord],
-) -> Result<()> {
-    if rows.is_empty() {
-        return Ok(());
-    }
-
-    let mut deltas: BTreeMap<(String, i64), ForwardProxyAttemptHourlyDelta> = BTreeMap::new();
-    for row in rows {
-        let bucket_start_epoch = forward_proxy_attempt_bucket_start_epoch(&row.occurred_at)?;
-        let entry = deltas
-            .entry((row.proxy_key.clone(), bucket_start_epoch))
-            .or_default();
-        entry.attempts += 1;
-        if row.is_success != 0 {
-            entry.success_count += 1;
-        } else {
-            entry.failure_count += 1;
-        }
-        if let Some(latency_ms) = row.latency_ms
-            && latency_ms.is_finite()
-            && latency_ms >= 0.0
-        {
-            entry.latency_sample_count += 1;
-            entry.latency_sum_ms += latency_ms;
-            entry.latency_max_ms = entry.latency_max_ms.max(latency_ms);
-        }
-    }
-
-    for ((proxy_key, bucket_start_epoch), delta) in deltas {
-        sqlx::query(
-            r#"
-            INSERT INTO forward_proxy_attempt_hourly (
-                proxy_key,
-                bucket_start_epoch,
-                attempts,
-                success_count,
-                failure_count,
-                latency_sample_count,
-                latency_sum_ms,
-                latency_max_ms,
-                updated_at
-            )
-            VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8, datetime('now'))
-            ON CONFLICT(proxy_key, bucket_start_epoch) DO UPDATE SET
-                attempts = forward_proxy_attempt_hourly.attempts + excluded.attempts,
-                success_count = forward_proxy_attempt_hourly.success_count + excluded.success_count,
-                failure_count = forward_proxy_attempt_hourly.failure_count + excluded.failure_count,
-                latency_sample_count = forward_proxy_attempt_hourly.latency_sample_count + excluded.latency_sample_count,
-                latency_sum_ms = forward_proxy_attempt_hourly.latency_sum_ms + excluded.latency_sum_ms,
-                latency_max_ms = MAX(forward_proxy_attempt_hourly.latency_max_ms, excluded.latency_max_ms),
-                updated_at = datetime('now')
-            "#,
-        )
-        .bind(&proxy_key)
-        .bind(bucket_start_epoch)
-        .bind(delta.attempts)
-        .bind(delta.success_count)
-        .bind(delta.failure_count)
-        .bind(delta.latency_sample_count)
-        .bind(delta.latency_sum_ms)
-        .bind(delta.latency_max_ms)
-        .execute(&mut *tx)
-        .await?;
-    }
-
-    Ok(())
-}
-
-async fn delete_hourly_rollup_rows_for_bucket_epochs_tx(
-    tx: &mut SqliteConnection,
-    table: &str,
-    bucket_epochs: &[i64],
-) -> Result<()> {
-    if bucket_epochs.is_empty() {
-        return Ok(());
-    }
-    let mut query =
-        QueryBuilder::<Sqlite>::new(format!("DELETE FROM {table} WHERE bucket_start_epoch IN ("));
-    {
-        let mut separated = query.separated(", ");
-        for bucket_epoch in bucket_epochs {
-            separated.push_bind(bucket_epoch);
-        }
-    }
-    query.push(")");
-    query.build().execute(&mut *tx).await?;
-    Ok(())
-}
-
-async fn load_live_invocation_hourly_rows_for_bucket_epochs_tx(
-    tx: &mut SqliteConnection,
-    bucket_epochs: &[i64],
-) -> Result<Vec<InvocationHourlySourceRecord>> {
-    if bucket_epochs.is_empty() {
-        return Ok(Vec::new());
-    }
-
-    let min_bucket_epoch = *bucket_epochs
-        .iter()
-        .min()
-        .ok_or_else(|| anyhow!("missing minimum invocation bucket epoch"))?;
-    let max_bucket_epoch = *bucket_epochs
-        .iter()
-        .max()
-        .ok_or_else(|| anyhow!("missing maximum invocation bucket epoch"))?;
-    let min_bucket_start = Utc
-        .timestamp_opt(min_bucket_epoch, 0)
-        .single()
-        .ok_or_else(|| anyhow!("invalid minimum invocation bucket epoch"))?;
-    let max_bucket_end = Utc
-        .timestamp_opt(max_bucket_epoch + 3_600, 0)
-        .single()
-        .ok_or_else(|| anyhow!("invalid maximum invocation bucket epoch"))?;
-    let bucket_epoch_set = bucket_epochs.iter().copied().collect::<HashSet<_>>();
-
-    let rows = sqlx::query_as::<_, InvocationHourlySourceRecord>(
-        "SELECT \
-            id,
-            occurred_at,
-            source,
-            status,
-            detail_level,
-            total_tokens,
-            cost,
-            error_message,
-            failure_kind,
-            failure_class,
-            is_actionable,
-            payload,
-            t_total_ms,
-            t_req_read_ms,
-            t_req_parse_ms,
-            t_upstream_connect_ms,
-            t_upstream_ttfb_ms,
-            t_upstream_stream_ms,
-            t_resp_parse_ms,
-            t_persist_ms
-         FROM codex_invocations
-         WHERE occurred_at >= ?1
-           AND occurred_at < ?2
-         ORDER BY id ASC",
-    )
-    .bind(db_occurred_at_lower_bound(min_bucket_start))
-    .bind(db_occurred_at_lower_bound(max_bucket_end))
-    .fetch_all(&mut *tx)
-    .await?;
-    Ok(rows
-        .into_iter()
-        .filter(|row| {
-            invocation_bucket_start_epoch(&row.occurred_at)
-                .map(|bucket_epoch| bucket_epoch_set.contains(&bucket_epoch))
-                .unwrap_or(false)
-        })
-        .collect())
-}
-
-async fn recompute_invocation_hourly_rollups_for_ids_tx(
-    tx: &mut SqliteConnection,
-    ids: &[i64],
-) -> Result<()> {
-    if ids.is_empty() {
-        return Ok(());
-    }
-
-    let mut query = QueryBuilder::<Sqlite>::new(
-        "SELECT DISTINCT occurred_at FROM codex_invocations WHERE id IN (",
-    );
-    {
-        let mut separated = query.separated(", ");
-        for id in ids {
-            separated.push_bind(id);
-        }
-    }
-    query.push(")");
-    let occurred_rows = query
-        .build_query_scalar::<String>()
-        .fetch_all(&mut *tx)
-        .await?;
-    if occurred_rows.is_empty() {
-        return Ok(());
-    }
-
-    let mut bucket_epochs = occurred_rows
-        .iter()
-        .map(|occurred_at| invocation_bucket_start_epoch(occurred_at))
-        .collect::<Result<Vec<_>>>()?;
-    bucket_epochs.sort_unstable();
-    bucket_epochs.dedup();
-    if bucket_epochs.is_empty() {
-        return Ok(());
-    }
-
-    for table in [
-        HOURLY_ROLLUP_TARGET_INVOCATIONS,
-        HOURLY_ROLLUP_TARGET_INVOCATION_FAILURES,
-        HOURLY_ROLLUP_TARGET_PROXY_PERF,
-        HOURLY_ROLLUP_TARGET_PROMPT_CACHE,
-        HOURLY_ROLLUP_TARGET_PROMPT_CACHE_UPSTREAM_ACCOUNTS,
-        HOURLY_ROLLUP_TARGET_STICKY_KEYS,
-    ] {
-        delete_hourly_rollup_rows_for_bucket_epochs_tx(tx, table, &bucket_epochs).await?;
-    }
-
-    let rows = load_live_invocation_hourly_rows_for_bucket_epochs_tx(tx, &bucket_epochs).await?;
-    upsert_invocation_hourly_rollups_tx(tx, &rows, &INVOCATION_HOURLY_ROLLUP_TARGETS).await?;
-    Ok(())
-}
-
-async fn replay_live_invocation_hourly_rollups(pool: &Pool<Sqlite>) -> Result<u64> {
-    let cursor_id =
-        load_hourly_rollup_live_progress(pool, HOURLY_ROLLUP_DATASET_INVOCATIONS).await?;
-    let rows = sqlx::query_as::<_, InvocationHourlySourceRecord>(
-        r#"
-        SELECT
-            id,
-            occurred_at,
-            source,
-            status,
-            detail_level,
-            total_tokens,
-            cost,
-            error_message,
-            failure_kind,
-            failure_class,
-            is_actionable,
-            payload,
-            t_total_ms,
-            t_req_read_ms,
-            t_req_parse_ms,
-            t_upstream_connect_ms,
-            t_upstream_ttfb_ms,
-            t_upstream_stream_ms,
-            t_resp_parse_ms,
-            t_persist_ms
-        FROM codex_invocations
-        WHERE id > ?1
-        ORDER BY id ASC
-        LIMIT ?2
-        "#,
-    )
-    .bind(cursor_id)
-    .bind(BACKFILL_BATCH_SIZE)
-    .fetch_all(pool)
-    .await?;
-    if rows.is_empty() {
-        return Ok(0);
-    }
-
-    let last_id = rows.last().map(|row| row.id).unwrap_or(cursor_id);
-    let mut tx = pool.begin().await?;
-    upsert_invocation_hourly_rollups_tx(tx.as_mut(), &rows, &INVOCATION_HOURLY_ROLLUP_TARGETS)
-        .await?;
-    save_hourly_rollup_live_progress_tx(tx.as_mut(), HOURLY_ROLLUP_DATASET_INVOCATIONS, last_id)
-        .await?;
-    tx.commit().await?;
-    Ok(rows.len() as u64)
-}
-
-async fn replay_live_invocation_hourly_rollups_tx(tx: &mut SqliteConnection) -> Result<u64> {
-    let cursor_id =
-        load_hourly_rollup_live_progress_tx(tx, HOURLY_ROLLUP_DATASET_INVOCATIONS).await?;
-    let rows = sqlx::query_as::<_, InvocationHourlySourceRecord>(
-        r#"
-        SELECT
-            id,
-            occurred_at,
-            source,
-            status,
-            detail_level,
-            total_tokens,
-            cost,
-            error_message,
-            failure_kind,
-            failure_class,
-            is_actionable,
-            payload,
-            t_total_ms,
-            t_req_read_ms,
-            t_req_parse_ms,
-            t_upstream_connect_ms,
-            t_upstream_ttfb_ms,
-            t_upstream_stream_ms,
-            t_resp_parse_ms,
-            t_persist_ms
-        FROM codex_invocations
-        WHERE id > ?1
-        ORDER BY id ASC
-        LIMIT ?2
-        "#,
-    )
-    .bind(cursor_id)
-    .bind(BACKFILL_BATCH_SIZE)
-    .fetch_all(&mut *tx)
-    .await?;
-    if rows.is_empty() {
-        return Ok(0);
-    }
-
-    let last_id = rows.last().map(|row| row.id).unwrap_or(cursor_id);
-    upsert_invocation_hourly_rollups_tx(tx, &rows, &INVOCATION_HOURLY_ROLLUP_TARGETS).await?;
-    save_hourly_rollup_live_progress_tx(tx, HOURLY_ROLLUP_DATASET_INVOCATIONS, last_id).await?;
-    Ok(rows.len() as u64)
-}
-
-async fn replay_live_forward_proxy_attempt_hourly_rollups(pool: &Pool<Sqlite>) -> Result<u64> {
-    let cursor_id =
-        load_hourly_rollup_live_progress(pool, HOURLY_ROLLUP_DATASET_FORWARD_PROXY_ATTEMPTS)
-            .await?;
-    let rows = sqlx::query_as::<_, ForwardProxyAttemptHourlySourceRecord>(
-        r#"
-        SELECT
-            id,
-            proxy_key,
-            occurred_at,
-            is_success,
-            latency_ms
-        FROM forward_proxy_attempts
-        WHERE id > ?1
-        ORDER BY id ASC
-        LIMIT ?2
-        "#,
-    )
-    .bind(cursor_id)
-    .bind(BACKFILL_BATCH_SIZE)
-    .fetch_all(pool)
-    .await?;
-    if rows.is_empty() {
-        return Ok(0);
-    }
-
-    let last_id = rows.last().map(|row| row.id).unwrap_or(cursor_id);
-    let mut tx = pool.begin().await?;
-    upsert_forward_proxy_attempt_hourly_rollups_tx(tx.as_mut(), &rows).await?;
-    save_hourly_rollup_live_progress_tx(
-        tx.as_mut(),
-        HOURLY_ROLLUP_DATASET_FORWARD_PROXY_ATTEMPTS,
-        last_id,
-    )
-    .await?;
-    tx.commit().await?;
-    Ok(rows.len() as u64)
-}
-
-async fn replay_live_forward_proxy_attempt_hourly_rollups_tx(
-    tx: &mut SqliteConnection,
-) -> Result<u64> {
-    let cursor_id =
-        load_hourly_rollup_live_progress_tx(tx, HOURLY_ROLLUP_DATASET_FORWARD_PROXY_ATTEMPTS)
-            .await?;
-    let rows = sqlx::query_as::<_, ForwardProxyAttemptHourlySourceRecord>(
-        r#"
-        SELECT
-            id,
-            proxy_key,
-            occurred_at,
-            is_success,
-            latency_ms
-        FROM forward_proxy_attempts
-        WHERE id > ?1
-        ORDER BY id ASC
-        LIMIT ?2
-        "#,
-    )
-    .bind(cursor_id)
-    .bind(BACKFILL_BATCH_SIZE)
-    .fetch_all(&mut *tx)
-    .await?;
-    if rows.is_empty() {
-        return Ok(0);
-    }
-
-    let last_id = rows.last().map(|row| row.id).unwrap_or(cursor_id);
-    upsert_forward_proxy_attempt_hourly_rollups_tx(tx, &rows).await?;
-    save_hourly_rollup_live_progress_tx(tx, HOURLY_ROLLUP_DATASET_FORWARD_PROXY_ATTEMPTS, last_id)
-        .await?;
-    Ok(rows.len() as u64)
-}
-
-async fn backfill_invocation_rollup_hourly_from_sources(pool: &Pool<Sqlite>) -> Result<usize> {
-    let archive_files = sqlx::query_as::<_, ArchiveBatchFileRow>(
-        r#"
-        SELECT id, file_path, coverage_start_at, coverage_end_at
-        FROM archive_batches
-        WHERE dataset = 'codex_invocations'
-          AND status = ?1
-        ORDER BY month_key ASC, created_at ASC, id ASC
-        "#,
-    )
-    .bind(ARCHIVE_STATUS_COMPLETED)
-    .fetch_all(pool)
-    .await?;
-    let mut overall: BTreeMap<(i64, String), InvocationHourlyRollupDelta> = BTreeMap::new();
-    let mut seen_ids = HashSet::new();
-
-    for archive_file in archive_files {
-        let archive_path = PathBuf::from(&archive_file.file_path);
-        if !archive_path.exists() {
-            warn!(
-                dataset = HOURLY_ROLLUP_DATASET_INVOCATIONS,
-                file_path = archive_file.file_path,
-                "skipping missing archive batch during invocation hourly rollup backfill"
-            );
-            continue;
-        }
-
-        let temp_path = PathBuf::from(format!(
-            "{}.{}.sqlite",
-            archive_path.display(),
-            retention_temp_suffix()
-        ));
-        if temp_path.exists() {
-            let _ = fs::remove_file(&temp_path);
-        }
-        let temp_cleanup = TempSqliteCleanup(temp_path.clone());
-        inflate_gzip_sqlite_file(&archive_path, &temp_path)?;
-        let archive_pool = SqlitePoolOptions::new()
-            .max_connections(1)
-            .connect(&sqlite_url_for_path(&temp_path))
-            .await
-            .with_context(|| format!("failed to open archive batch {}", archive_path.display()))?;
-        let mut archive_cursor_id = 0_i64;
-        loop {
-            let mut rows = sqlx::query_as::<_, InvocationHourlySourceRecord>(
-                r#"
-                SELECT
-                    id,
-                    occurred_at,
-                    source,
-                    status,
-                    detail_level,
-                    total_tokens,
-                    cost,
-                    error_message,
-                    failure_kind,
-                    failure_class,
-                    is_actionable,
-                    payload,
-                    t_total_ms,
-                    t_req_read_ms,
-                    t_req_parse_ms,
-                    t_upstream_connect_ms,
-                    t_upstream_ttfb_ms,
-                    t_upstream_stream_ms,
-                    t_resp_parse_ms,
-                    t_persist_ms
-                FROM codex_invocations
-                WHERE id > ?1
-                ORDER BY id ASC
-                LIMIT ?2
-                "#,
-            )
-            .bind(archive_cursor_id)
-            .bind(BACKFILL_BATCH_SIZE)
-            .fetch_all(&archive_pool)
-            .await?;
-            if rows.is_empty() {
-                break;
-            }
-            archive_cursor_id = rows.last().map(|row| row.id).unwrap_or(archive_cursor_id);
-            rows.retain(|row| seen_ids.insert(row.id));
-            if rows.is_empty() {
-                continue;
-            }
-            accumulate_invocation_hourly_overall_rollups(&mut overall, &rows)?;
-        }
-        archive_pool.close().await;
-        drop(temp_cleanup);
-    }
-
-    let mut cursor_id = 0_i64;
-    loop {
-        let mut rows = sqlx::query_as::<_, InvocationHourlySourceRecord>(
-            r#"
-            SELECT
-                id,
-                occurred_at,
-                source,
-                status,
-                detail_level,
-                total_tokens,
-                cost,
-                error_message,
-                failure_kind,
-                failure_class,
-                is_actionable,
-                payload,
-                t_total_ms,
-                t_req_read_ms,
-                t_req_parse_ms,
-                t_upstream_connect_ms,
-                t_upstream_ttfb_ms,
-                t_upstream_stream_ms,
-                t_resp_parse_ms,
-                t_persist_ms
-            FROM codex_invocations
-            WHERE id > ?1
-            ORDER BY id ASC
-            LIMIT ?2
-            "#,
-        )
-        .bind(cursor_id)
-        .bind(BACKFILL_BATCH_SIZE)
-        .fetch_all(pool)
-        .await?;
-        if rows.is_empty() {
-            break;
-        }
-        cursor_id = rows.last().map(|row| row.id).unwrap_or(cursor_id);
-        rows.retain(|row| seen_ids.insert(row.id));
-        if rows.is_empty() {
-            continue;
-        }
-        accumulate_invocation_hourly_overall_rollups(&mut overall, &rows)?;
-    }
-
-    if overall.is_empty() {
-        return Ok(0);
-    }
-
-    let mut tx = pool.begin().await?;
-    for ((bucket_start_epoch, source), delta) in &overall {
-        sqlx::query(
-            r#"
-            INSERT INTO invocation_rollup_hourly (
-                bucket_start_epoch,
-                source,
-                total_count,
-                success_count,
-                failure_count,
-                total_tokens,
-                total_cost,
-                first_byte_sample_count,
-                first_byte_sum_ms,
-                first_byte_max_ms,
-                first_byte_histogram,
-                first_response_byte_total_sample_count,
-                first_response_byte_total_sum_ms,
-                first_response_byte_total_max_ms,
-                first_response_byte_total_histogram,
-                updated_at
-            )
-            VALUES (?1, ?2, ?3, ?4, ?5, ?6, ?7, ?8, ?9, ?10, ?11, ?12, ?13, ?14, ?15, datetime('now'))
-            ON CONFLICT(bucket_start_epoch, source) DO UPDATE SET
-                total_count = excluded.total_count,
-                success_count = excluded.success_count,
-                failure_count = excluded.failure_count,
-                total_tokens = excluded.total_tokens,
-                total_cost = excluded.total_cost,
-                first_byte_sample_count = excluded.first_byte_sample_count,
-                first_byte_sum_ms = excluded.first_byte_sum_ms,
-                first_byte_max_ms = excluded.first_byte_max_ms,
-                first_byte_histogram = excluded.first_byte_histogram,
-                first_response_byte_total_sample_count = excluded.first_response_byte_total_sample_count,
-                first_response_byte_total_sum_ms = excluded.first_response_byte_total_sum_ms,
-                first_response_byte_total_max_ms = excluded.first_response_byte_total_max_ms,
-                first_response_byte_total_histogram = excluded.first_response_byte_total_histogram,
-                updated_at = datetime('now')
-            "#,
-        )
-        .bind(*bucket_start_epoch)
-        .bind(source)
-        .bind(delta.total_count)
-        .bind(delta.success_count)
-        .bind(delta.failure_count)
-        .bind(delta.total_tokens)
-        .bind(delta.total_cost)
-        .bind(delta.first_byte_sample_count)
-        .bind(delta.first_byte_sum_ms)
-        .bind(delta.first_byte_max_ms)
-        .bind(encode_approx_histogram(&delta.first_byte_histogram)?)
-        .bind(delta.first_response_byte_total_sample_count)
-        .bind(delta.first_response_byte_total_sum_ms)
-        .bind(delta.first_response_byte_total_max_ms)
-        .bind(encode_approx_histogram(
-            &delta.first_response_byte_total_histogram,
-        )?)
-        .execute(tx.as_mut())
-        .await?;
-    }
-    tx.commit().await?;
-
-    Ok(overall.len())
-}
+#[path = "archive/cleanup.rs"]
+mod archive_cleanup;
+#[path = "archive/manifest.rs"]
+mod archive_manifest;
+#[path = "archive/writers.rs"]
+mod archive_writers;
+#[path = "archive/hourly_rollups.rs"]
+mod archive_hourly_rollups;
+
+pub(crate) use archive_cleanup::*;
+pub(crate) use archive_manifest::*;
+pub(crate) use archive_writers::*;
+pub(crate) use archive_hourly_rollups::*;
