+++ src/maintenance/archive/hourly_rollup_support.rs
